- 修改：拆分桌面端应用状态、设置存储与各领域 Tauri API 适配层
- 修改：会话同步和删除时安全关闭并恢复 Codex Desktop

- 新增：Skills 与 MCP 备注持久化及编辑界面

- 新增：桌面端架构、数据边界与维护流程文档

- 修改：补充会话生命周期、备注功能及数据库迁移测试
