+++ apps/desktop/src/app/App.tsx
@@ -1,0 +1,3153 @@
+import React from "react";
+import { flushSync } from "react-dom";
+import { Loader2 } from "lucide-react";
+import { SessionManagementPage } from "../pages/SessionManagementPage";
+import type { SessionPreview, SessionSyncStatus } from "../features/sessions/types";
+import { OverviewPage } from "../pages/OverviewPage";
+import { AboutPage, SettingsPage, TomlConfigPage } from "../pages/UtilityPages";
+import { PromptsPage } from "../pages/PromptsPage";
+import { SkillsMcpPage } from "../pages/SkillsMcpPage";
+import { ProvidersPage, type ProviderCopy, type ProviderRow } from "../pages/ProvidersPage";
+import { AppShell, type AppTab, type AppTheme } from "../components/AppShell";
+import {
+  AppToast,
+  StartupWizardDialog,
+  UpdateDialog,
+} from "../components/AppDialogs";
+import { PageTransition } from "../components/PageTransition";
+import { cx } from "../components/ui";
+import { appUpdater, useAppUpdater } from "../appUpdater";
+import * as codexApi from "../features/codex/api";
+import * as officialAccountsApi from "../features/official-accounts/api";
+import * as promptsApi from "../features/prompts/api";
+import * as providersApi from "../features/providers/api";
+import * as sessionsApi from "../features/sessions/api";
+import * as skillsMcpApi from "../features/skills-mcp/api";
+import * as updatesApi from "../features/updates/api";
+import {
+  clearStartupWizardSeen,
+  getActiveProviderId,
+  getAutoRestartCodex,
+  getLanguage,
+  getPromptInjectionMode,
+  getTheme,
+  hasSeenStartupWizard,
+  markStartupWizardSeen,
+  setActiveProviderId as persistActiveProviderId,
+  setAutoRestartCodex as persistAutoRestartCodex,
+  setLanguage as persistLanguage,
+  setPromptInjectionMode as persistPromptInjectionMode,
+  setTheme as persistTheme,
+} from "./appStorage";
+import type {
+  AboutInfo,
+  ActionResult,
+  AppUpdateInfo,
+  BuiltinPromptDetail,
+  BuiltinPromptStatus,
+  CodexRestartResult,
+  CodexState,
+  ImportResult,
+  InstructionMode,
+  InstructionTemplate,
+  Lang,
+  OfficialAccountDraft,
+  OfficialAccountSummary,
+  OfficialAuthCandidate,
+  OfficialConfigDraft,
+  PromptInjectionMode,
+  ProviderConnectionResult,
+  ProviderModel,
+  ProviderModelsResult,
+  ProviderMode,
+  ReleaseInfo,
+  SavedPrompt,
+  SavedProvider,
+  SessionDeleteResult,
+  SessionSyncResult,
+  SkillsMcpActionResult,
+  SkillsMcpImportPreview,
+  SkillsMcpState,
+  StartupDiagnostics,
+} from "../types";
+import "../styles/base.css";
+import "../styles/app-shell.css";
+import "../styles/ui-primitives.css";
+import "../styles/app-dialogs.css";
+import "../styles/dark-theme.css";
+
+type Tab = AppTab;
+
+const FALLBACK_GITHUB_REPO = "yynxxxxx/Codex-X";
+
+type ThemeTransitionDocument = Document & {
+  startViewTransition?: (update: () => void | Promise<void>) => { finished: Promise<void> };
+};
+
+const bundledInstructionTemplates: InstructionTemplate[] = [
+  {
+    id: "gpt5.5-unrestricted",
+    filename: "gpt5.5-unrestricted.md",
+    title: "gpt-5.5 unrestricted 破甲",
+    subtitle: "方法：先让ai分析项目，分析完之后发【不直白的逆向】命令",
+    badge: "推荐",
+  },
+  {
+    id: "gpt5.4-unrestricted",
+    filename: "gpt5.4-unrestricted.md",
+    title: "gpt-5.4 unrestricted 破甲",
+    subtitle: "方法：先让ai分析项目，分析完之后发【不直白的逆向】命令",
+    badge: "兼容",
+  },
+  {
+    id: "gpt5.5-jeli",
+    filename: "gpt5.5-jeli.md",
+    title: "gpt5.5-jeli.md",
+    subtitle: "gpt5.5 大白话（80%场景）破甲",
+    badge: "通用",
+  },
+  {
+    id: "github-gpt-5-6-sol-unrestricted-33b86c71",
+    filename: "gpt-5.6-sol-unrestricted.md",
+    title: "gpt-5.6-sol-unrestricted.md",
+    subtitle: "gpt5.6-sol 破甲提示词",
+    badge: "内置",
+  },
+  {
+    id: "github-3-0-b459e1e8",
+    filename: "海鸥3.0破甲.md",
+    title: "海鸥3.0破甲.md",
+    subtitle: "测试生效：海鸥在线，你要整点薯条吗？",
+    badge: "内置",
+  },
+];
+
+const defaultProviderForm: SavedProvider = {
+  id: "magicai",
+  providerName: "MagicAI",
+  baseUrl: "https://sky1818.com",
+  model: "gpt-5.5",
+  apiKey: "",
+  tomlConfig: "",
+  wireApi: "responses",
+  requiresOpenaiAuth: false,
+};
+
+const blankProviderForm: SavedProvider = {
+  id: "",
+  providerName: "",
+  baseUrl: "",
+  model: "gpt-5.5",
+  apiKey: "",
+  tomlConfig: "",
+  wireApi: "responses",
+  requiresOpenaiAuth: false,
+};
+
+const blankPromptForm: SavedPrompt = {
+  id: "",
+  title: "",
+  filename: "",
+  content: "",
+};
+
+const dict = {
+  zh: {
+    appSubtitle: "切换 · 指令 · 配置",
+    manager: "Codex 配置管理器",
+    load: "加载",
+    refresh: "刷新",
+    nav: {
+      dashboard: "概览",
+      provider: "供应商",
+      sessions: "会话管理",
+      skillsMcp: "技能和MCP",
+      instruction: "指令提示词",
+      toml: "TOML",
+      settings: "设置",
+      about: "关于",
+    },
+    dashboard: {
+      config: "配置文件",
+      found: "已找到",
+      missing: "不存在",
+      provider: "供应商",
+      instruction: "指令提示词状态",
+      enabled: "已启用",
+      disabled: "未启用",
+      auth: "认证文件",
+      currentConfig: "当前 Codex 配置",
+      liveStatus: "实时状态",
+      dir: "目录",
+      configPath: "配置",
+      model: "模型",
+      providerName: "供应商",
+      instructionFile: "指令文件",
+      notSet: "未设置",
+      officialDefault: "官方默认",
+    },
+    provider: {
+      title: "供应商列表",
+      subtitle: "像 cc-switch 一样管理 Codex 第三方 API。点击卡片可切换，点击 + 添加新供应商。",
+      add: "添加供应商",
+      importCc: "从 cc-switch 导入",
+      edit: "编辑",
+      viewEdit: "编辑",
+      remove: "删除",
+      switch: "切换",
+      current: "当前",
+      official: "官方配置",
+      noRouting: "不支持路由",
+      authReady: "认证文件存在",
+      authMissing: "未找到认证文件",
+      detected: "从 TOML 检测",
+      local: "本地保存",
+      noProviders: "还没有供应商，点击右上角 + 添加。",
+      officialEdit: "OpenAI Official 编辑",
+      officialHint: "官方配置与当前中转配置独立保存。读取配置不会切换当前供应商；切换回官方时会优先使用已保存配置。",
+      officialUrl: "官方入口",
+      formAdd: "添加新供应商",
+      formEdit: "编辑供应商",
+      formHint: "保存后会写入供应商列表，并同步写入 Codex live 配置。下方可预览将生成的 config.toml。",
+      name: "供应商名称",
+      baseUrl: "Base URL",
+      model: "模型",
+      wireApi: "Wire API",
+      apiKey: "API Key",
+      apiKeyPlaceholder: "用于写入 auth.json",
+      requiresAuth: "requires_openai_auth",
+      save: "保存到列表",
+      saveAndSwitch: "保存",
+      cancel: "返回列表",
+    },
+    instruction: {
+      title: "一键管理指令提示词",
+      desc: "启用时写入指令提示词文件并设置 model_instructions_file；禁用时只移除 Codex-X 管理的指令提示词字段并删除 md 文件。每次操作前都会创建备份。",
+      enabled: "已启用",
+      disabled: "未启用",
+      unset: "model_instructions_file 未设置",
+      enable: "启用",
+      disable: "禁用 / 删除",
+    },
+    toml: {
+      title: "当前 live TOML 配置",
+      desc: "这里显示的是 Codex 当前正在使用的 ~/.codex/config.toml，不是本地保存的供应商模板。切换供应商后，这里会变成新写入的 live 配置。",
+      loaded: "已读取",
+      missingText: "# config.toml 不存在，执行切换或启用后会自动创建。",
+    },
+    backups: {
+      title: "备份与撤回",
+      empty: "还没有备份。首次写入前会自动创建。",
+      restore: "恢复",
+    },
+    settings: {
+      title: "设置",
+      language: "界面语言",
+      zh: "中文",
+      en: "English",
+      languageDesc: "默认中文，可随时切换。设置会保存在浏览器本地存储。",
+      productName: "产品名",
+      productDesc: "当前名称为 Codex-X，定位是 Codex Switch & Instruct。",
+    },
+    loadingConfig: "正在读取 Codex 配置...",
+    noAuth: "无 auth",
+    authJson: "auth.json",
+  },
+  en: {
+    appSubtitle: "Switch · Instruct · Config",
+    manager: "Codex config manager",
+    load: "Load",
+    refresh: "Refresh",
+    nav: {
+      dashboard: "Overview",
+      provider: "Provider",
+      sessions: "Sessions",
+      skillsMcp: "Skills & MCP",
+      instruction: "Prompt",
+      toml: "TOML",
+      settings: "Settings",
+      about: "About",
+    },
+    dashboard: {
+      config: "Config",
+      found: "Found",
+      missing: "Missing",
+      provider: "Provider",
+      instruction: "Instruction Prompt",
+      enabled: "Enabled",
+      disabled: "Disabled",
+      auth: "Auth",
+      currentConfig: "Current Codex config",
+      liveStatus: "Live status",
+      dir: "Directory",
+      configPath: "Config",
+      model: "Model",
+      providerName: "Provider",
+      instructionFile: "Instruction",
+      notSet: "Not set",
+      officialDefault: "Official / Default",
+    },
+    provider: {
+      title: "Provider list",
+      subtitle: "Manage Codex third-party APIs like cc-switch. Click a row to switch; use + to add a provider.",
+      add: "Add provider",
+      importCc: "Import from cc-switch",
+      edit: "Edit",
+      viewEdit: "Edit",
+      remove: "Delete",
+      switch: "Switch",
+      current: "Current",
+      official: "Official",
+      noRouting: "No routing",
+      authReady: "Auth found",
+      authMissing: "Auth missing",
+      detected: "Detected from TOML",
+      local: "Local",
+      noProviders: "No provider yet. Click + to add one.",
+      officialEdit: "OpenAI Official settings",
+      officialHint: "Official settings are stored separately from the active proxy. Loading a config does not switch providers; switching back uses the saved config first.",
+      officialUrl: "Official URL",
+      formAdd: "Add provider",
+      formEdit: "Edit provider",
+      formHint: "Save writes the provider to the list and applies it to the Codex live config. The generated config.toml is previewed below.",
+      name: "Provider name",
+      baseUrl: "Base URL",
+      model: "Model",
+      wireApi: "Wire API",
+      apiKey: "API Key",
+      apiKeyPlaceholder: "Written to auth.json",
+      requiresAuth: "requires_openai_auth",
+      save: "Save",
+      saveAndSwitch: "Save",
+      cancel: "Back",
+    },
+    instruction: {
+      title: "Manage instruction prompt",
+      desc: "Enable writes the instruction prompt file and sets model_instructions_file; disable removes Codex-X-managed instruction prompt config and deletes the md file. Every write creates a backup first.",
+      enabled: "Enabled",
+      disabled: "Disabled",
+      unset: "model_instructions_file is not set",
+      enable: "Enable",
+      disable: "Disable / delete",
+    },
+    toml: {
+      title: "Current live TOML config",
+      desc: "This is the active ~/.codex/config.toml used by Codex, not a saved provider template. After switching providers, this page shows the newly written live config.",
+      loaded: "Loaded",
+      missingText: "# config.toml is missing. It will be created after switching or enabling instruction.",
+    },
+    backups: {
+      title: "Backups & restore",
+      empty: "No backups yet. A backup will be created before the first write.",
+      restore: "Restore",
+    },
+    settings: {
+      title: "Settings",
+      language: "Language",
+      zh: "中文",
+      en: "English",
+      languageDesc: "Chinese is the default. You can switch at any time; the setting is saved locally.",
+      productName: "Product name",
+      productDesc: "Current name is Codex-X, positioned as Codex Switch & Instruct.",
+    },
+    loadingConfig: "Reading Codex config...",
+    noAuth: "No auth",
+    authJson: "auth.json",
+  },
+} as const;
+
+function getProviderPageCopy(lang: Lang): ProviderCopy {
+  const t = dict[lang];
+  const isChinese = lang === "zh";
+  return {
+    eyebrow: "Provider",
+    title: t.provider.title,
+    subtitle: t.provider.subtitle,
+    importLabel: t.provider.importCc,
+    restartCodexLabel: isChinese ? "重启 Codex" : "Restart Codex",
+    restartingCodexLabel: isChinese ? "正在重启..." : "Restarting...",
+    addLabel: t.provider.add,
+    addOfficialAccountLabel: isChinese ? "新增官方账号" : "Add official account",
+    saveCurrentOfficialLabel: isChinese ? "保存当前登录" : "Save current login",
+    captureOfficialTitle: isChinese ? "保存当前 OpenAI 登录" : "Save current OpenAI login",
+    captureOfficialDescription: isChinese
+      ? "为当前登录设置一个名称。认证内容只保存在本机，不会显示在账号列表中。"
+      : "Name the current login. Credentials remain on this device and are never shown in the account list.",
+    captureOfficialNameLabel: isChinese ? "账号名称" : "Account name",
+    captureOfficialConfirmLabel: isChinese ? "保存登录" : "Save login",
+    noProviders: t.provider.noProviders,
+    currentLabel: isChinese ? "当前使用" : "Current",
+    enableLabel: isChinese ? "启用" : "Enable",
+    testLabel: isChinese ? "测试连接" : "Test connection",
+    editLabel: t.provider.edit,
+    copyLabel: isChinese ? "复制供应商" : "Copy provider",
+    removeLabel: t.provider.remove,
+    deleteTitle: isChinese ? "删除供应商" : "Delete provider",
+    deleteDescription: (providerName) => isChinese
+      ? `“${providerName}”将从供应商列表中删除，此操作无法撤销。`
+      : `“${providerName}” will be removed from the provider list. This cannot be undone.`,
+    deleteCurrentDescription: (providerName) => isChinese
+      ? `“${providerName}”当前正在使用。删除前会先热切换到 OpenAI Official，确定继续吗？`
+      : `“${providerName}” is currently active. Codex-X will hot-switch to OpenAI Official before deleting it. Continue?`,
+    deleteCurrentOfficialDescription: isChinese
+      ? "删除当前官方账号会清除当前 OpenAI 登录状态，需要重新登录或切换其他已保存账号。"
+      : "Deleting the current official account clears the current OpenAI login. You will need to sign in again or switch to another saved account.",
+    deleteCancelLabel: isChinese ? "取消" : "Cancel",
+    deleteConfirmLabel: isChinese ? "确认删除" : "Delete",
+    noBaseUrlLabel: "no base_url",
+    officialEyebrow: "OpenAI Official",
+    officialTitle: t.provider.officialEdit,
+    officialHint: t.provider.officialHint,
+    officialUrlLabel: t.provider.officialUrl,
+    authPathLabel: "auth.json",
+    officialCurrentLabel: t.provider.current,
+    officialAuthLabel: "auth.json (JSON)",
+    officialTomlLabel: "config.toml (TOML)",
+    officialSaveLabel: isChinese ? "保存官方配置" : "Save official config",
+    loadCcSwitchOfficialLabel: isChinese ? "从 CC Switch 载入" : "Load from CC Switch",
+    restoreOfficialLabel: isChinese ? "读取已保存配置" : "Load saved config",
+    resetOfficialLabel: isChinese ? "新建官方配置" : "Create official config",
+    resetOfficialTitle: isChinese ? "新建官方配置" : "Create official config",
+    resetOfficialDescription: isChinese
+      ? "这会切换到 OpenAI Official、清除当前 live auth.json，并要求你在 Codex 中重新登录。操作前会自动备份当前配置。"
+      : "This switches to OpenAI Official, removes the live auth.json, and requires a new Codex login. The current files are backed up first.",
+    resetOfficialCancelLabel: isChinese ? "取消" : "Cancel",
+    resetOfficialConfirmLabel: isChinese ? "清除并新建" : "Clear and create",
+    cancelLabel: t.provider.cancel,
+    formEyebrow: "Provider",
+    formAddTitle: t.provider.formAdd,
+    formEditTitle: t.provider.formEdit,
+    formHint: t.provider.formHint,
+    apiConfigTitle: isChinese ? "供应商 API 配置" : "Provider API configuration",
+    apiConfigDescription: isChinese
+      ? "在同一个页面管理 API、认证信息和 config.toml。"
+      : "Manage API, authentication, and config.toml in one place.",
+    apiKeyLabel: t.provider.apiKey,
+    apiKeyPlaceholder: t.provider.apiKeyPlaceholder,
+    showApiKeyLabel: isChinese ? "显示 API Key" : "Show API key",
+    hideApiKeyLabel: isChinese ? "隐藏 API Key" : "Hide API key",
+    baseUrlLabel: isChinese ? "API 请求地址" : t.provider.baseUrl,
+    nameLabel: t.provider.name,
+    modelLabel: t.provider.model,
+    accountNameLabel: isChinese ? "账号名称" : "Account name",
+    fetchModelsLabel: isChinese ? "获取模型列表" : "Fetch models",
+    fetchingModelsLabel: isChinese ? "获取中" : "Fetching",
+    chooseModelLabel: (count) => isChinese ? `选择已获取的模型（${count}）` : `Choose a fetched model (${count})`,
+    wireApiLabel: t.provider.wireApi,
+    requiresAuthLabel: t.provider.requiresAuth,
+    authPreviewTitle: "auth.json (JSON)",
+    authPreviewDescription: isChinese
+      ? "启用该供应商时，auth.json 只写入 OPENAI_API_KEY；API Key 留空会移除第三方认证文件。"
+      : "When enabled, auth.json contains only OPENAI_API_KEY. Leaving the API key empty removes third-party authentication.",
+    tomlTitle: "config.toml (TOML)",
+    tomlDescription: isChinese
+      ? "上方标准字段是启用时的权威值；已有模板中的其他扩展字段会保留。只有点击“重置生成”才会替换为标准模板。"
+      : "The standard fields above are authoritative when enabled. Other fields from an existing template are preserved; only Reset replaces it with the standard template.",
+    resetTomlLabel: isChinese ? "重置生成" : "Reset",
+    saveLabel: t.provider.saveAndSwitch,
+    savingLabel: isChinese ? "保存中..." : "Saving...",
+  };
+}
+function providerId(name: string) {
+  const slug = name
+    .trim()
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/^-+|-+$/g, "");
+  return slug || `provider-${Date.now()}`;
+}
+
+function isReservedCodexProviderId(id: string) {
+  return ["openai", "custom", "amazon-bedrock", "ollama", "lmstudio", "oss"].includes(id.trim().toLowerCase());
+}
+
+function customProviderId(name: string) {
+  const id = providerId(name);
+  return isReservedCodexProviderId(id) ? `${id}-custom` : id;
+}
+
+function uniqueId(base: string, existingIds: Iterable<string>) {
+  const used = new Set(Array.from(existingIds).map((id) => id.trim().toLowerCase()));
+  const clean = providerId(base);
+  let candidate = clean;
+  let index = 2;
+  while (used.has(candidate.toLowerCase())) {
+    candidate = `${clean}-${index}`;
+    index += 1;
+  }
+  return candidate;
+}
+
+function splitMarkdownFilename(filename: string) {
+  const clean = filename.trim().replace(/[\/\\]+/g, "-") || "prompt.md";
+  const stem = clean.replace(/\.md$/i, "") || "prompt";
+  return { stem, filename: `${stem}.md` };
+}
+
+function uniquePromptFilename(filename: string, existingFilenames: Iterable<string>) {
+  const used = new Set(Array.from(existingFilenames).map((name) => name.trim().toLowerCase()));
+  const { stem } = splitMarkdownFilename(filename);
+  let candidate = `${stem}.md`;
+  let index = 2;
+  while (used.has(candidate.toLowerCase())) {
+    candidate = `${stem}-${index}.md`;
+    index += 1;
+  }
+  return candidate;
+}
+
+function tomlEscape(value: string) {
+  return value.replace(/\\/g, "\\\\").replace(/"/g, '\\"');
+}
+
+function extractOpenAiApiKey(authText?: string) {
+  if (!authText?.trim()) return "";
+  try {
+    const parsed = JSON.parse(authText) as { OPENAI_API_KEY?: unknown };
+    return typeof parsed.OPENAI_API_KEY === "string" ? parsed.OPENAI_API_KEY : "";
+  } catch {
+    return "";
+  }
+}
+
+function normalizeProviderBaseUrl(value?: string | null) {
+  const raw = (value || "").trim();
+  if (!raw) return "";
+  try {
+    const parsed = new URL(raw);
+    const credentials = parsed.username
+      ? `${parsed.username}${parsed.password ? `:${parsed.password}` : ""}@`
+      : "";
+    const path = parsed.pathname.replace(/\/+$/, "");
+    return `${parsed.protocol.toLowerCase()}//${credentials}${parsed.host.toLowerCase()}${path}${parsed.search}`;
+  } catch {
+    return raw.replace(/\/+$/, "");
+  }
+}
+
+function normalizeProviderName(value?: string | null) {
+  return (value || "").trim().replace(/\s+/gu, " ").toLowerCase();
+}
+
+function parseTomlStringValue(value: string) {
+  const raw = value.trim();
+  if (raw.startsWith('"')) {
+    try {
+      return JSON.parse(raw) as string;
+    } catch {
+      const end = raw.lastIndexOf('"');
+      return end > 0 ? raw.slice(1, end) : raw.slice(1);
+    }
+  }
+  if (raw.startsWith("'")) {
+    const end = raw.lastIndexOf("'");
+    return end > 0 ? raw.slice(1, end) : raw.slice(1);
+  }
+  return raw.replace(/\s+#.*$/, "").trim();
+}
+
+function extractTomlProviderApiKey(configText: string | undefined, providerId?: string) {
+  if (!configText?.trim()) return "";
+  const targetSection = providerId ? `model_providers.${providerId}` : "";
+  let currentSection = "";
+  let topLevelValue = "";
+  let firstProviderValue = "";
+
+  for (const line of configText.replace(/\r\n?/g, "\n").split("\n")) {
+    const section = line.match(/^\s*\[([^\]]+)]\s*(?:#.*)?$/);
+    if (section) {
+      currentSection = section[1].trim();
+      continue;
+    }
+    const token = line.match(/^\s*experimental_bearer_token\s*=\s*(.+?)\s*$/);
+    if (!token) continue;
+    const value = parseTomlStringValue(token[1]).trim();
+    if (!value) continue;
+    if (!currentSection) topLevelValue = value;
+    if (currentSection.startsWith("model_providers.") && !firstProviderValue) firstProviderValue = value;
+    if (targetSection && currentSection === targetSection) return value;
+  }
+
+  return topLevelValue || (!providerId ? firstProviderValue : "");
+}
+
+function extractTomlModelProvider(configText: string | undefined) {
+  if (!configText?.trim()) return "";
+  for (const line of configText.replace(/\r\n?/g, "\n").split("\n")) {
+    if (/^\s*\[/.test(line)) break;
+    const entry = line.match(/^\s*model_provider\s*=\s*(.+?)\s*$/);
+    if (entry) return parseTomlStringValue(entry[1]).trim();
+  }
+  return "";
+}
+
+function savedProviderApiKey(provider: SavedProvider) {
+  const providerId = extractTomlModelProvider(provider.tomlConfig);
+  return (provider.apiKey || "").trim()
+    || extractTomlProviderApiKey(provider.tomlConfig, providerId || undefined);
+}
+
+function providerIdentityKey(
+  provider: Pick<SavedProvider, "baseUrl" | "providerName" | "model" | "wireApi" | "requiresOpenaiAuth">,
+  apiKey?: string | null,
+) {
+  const normalizedUrl = normalizeProviderBaseUrl(provider.baseUrl);
+  if (!normalizedUrl) return "";
+  const normalizedKey = (apiKey || "").trim();
+  return JSON.stringify([
+    normalizedUrl,
+    normalizedKey ? `key:${normalizedKey}` : `name:${normalizeProviderName(provider.providerName)}`,
+    provider.model.trim(),
+    provider.wireApi.trim().toLowerCase(),
+    provider.requiresOpenaiAuth,
+  ]);
+}
+
+function buildProviderTomlPreview(provider: SavedProvider) {
+  const model = provider.model.trim();
+  const name = provider.providerName.trim();
+  const providerKey = "custom";
+  const baseUrl = provider.baseUrl.trim().replace(/\/+$/, "");
+  if (!model || !name || !baseUrl) return "";
+  const wireApi = provider.wireApi || "responses";
+  return [
+    `model_provider = "${tomlEscape(providerKey)}"`,
+    `model = "${tomlEscape(model)}"`,
+    "",
+    `[model_providers.${providerKey}]`,
+    `name = "${tomlEscape(name)}"`,
+    `base_url = "${tomlEscape(baseUrl)}"`,
+    `wire_api = "${tomlEscape(wireApi)}"`,
+    `requires_openai_auth = ${provider.requiresOpenaiAuth ? "true" : "false"}`,
+  ].join("\n");
+}
+
+function buildOfficialTomlPreview(model: string) {
+  return [
+    `model_provider = "custom"`,
+    `model = "${tomlEscape(model.trim() || "gpt-5.5")}"`,
+    "",
+    "[model_providers.custom]",
+    `name = "OpenAI"`,
+    `wire_api = "responses"`,
+    `requires_openai_auth = true`,
+    `supports_websockets = true`,
+  ].join("\n");
+}
+
+function buildProviderAuthPreview(provider: SavedProvider) {
+  const key = provider.apiKey?.trim();
+  return JSON.stringify(key ? { OPENAI_API_KEY: key } : {}, null, 2);
+}
+
+function sessionMismatchCount(status: SessionSyncStatus | null) {
+  if (!status?.needsSync) return 0;
+  const exactCount = Number.isFinite(status.mismatchedSessions)
+    ? status.mismatchedSessions
+    : Math.max(status.mismatchedThreads, status.mismatchedRollouts);
+  return Math.max(1, exactCount);
+}
+
+function instructionIdFromPath(path: string | undefined, templates: InstructionTemplate[]) {
+  if (!path) return "";
+  const normalized = path.replace(/\\/g, "/");
+  const found = templates.find((item) => normalized.toLowerCase().endsWith(item.filename.toLowerCase()));
+  return found?.id || "custom";
+}
+
+function uniqueBuiltinPromptStatuses(statuses: BuiltinPromptStatus[]) {
+  const sourcePriority: Record<string, number> = {
+    unavailable: 0,
+    bundled: 1,
+    cache: 2,
+    removed: 2,
+    github: 3,
+  };
+  const seenIds = new Set<string>();
+  const seenFilenames = new Set<string>();
+  const selected = statuses
+    .map((item, index) => ({ item, index }))
+    .filter(({ item }) => item.id.trim() && item.filename.trim())
+    .sort((a, b) =>
+      (sourcePriority[b.item.contentSource] ?? -1) - (sourcePriority[a.item.contentSource] ?? -1)
+      || a.index - b.index,
+    )
+    .filter(({ item }) => {
+      const id = item.id.trim().toLowerCase();
+      const filename = item.filename.trim().toLowerCase();
+      if (seenIds.has(id) || seenFilenames.has(filename)) return false;
+      seenIds.add(id);
+      seenFilenames.add(filename);
+      return true;
+    });
+  return selected.sort((a, b) => a.index - b.index).map(({ item }) => item);
+}
+
+function JsonPreview({ text }: { text: string }) {
+  return (
+    <pre className="toml-preview json-preview" aria-label="JSON preview">
+      {text.split("\n").map((line, index) => (
+        <div className="toml-line" key={index}>
+          <span className="toml-line-no">{index + 1}</span>
+          <code>{line}</code>
+        </div>
+      ))}
+    </pre>
+  );
+}
+
+function renderTomlValue(value: string, lineKey: string) {
+  const parts = value.split(/("(?:\\.|[^"])*")/g);
+  return parts.map((part, index) => {
+    if (!part) return null;
+    const key = `${lineKey}-v-${index}`;
+    if (/^"(?:\\.|[^"])*"$/.test(part)) {
+      return <span className="toml-string" key={key}>{part}</span>;
+    }
+    const boolParts = part.split(/\b(true|false)\b/g);
+    return boolParts.map((piece, boolIndex) => {
+      if (piece === "true" || piece === "false") {
+        return <span className="toml-bool" key={`${key}-b-${boolIndex}`}>{piece}</span>;
+      }
+      return <React.Fragment key={`${key}-t-${boolIndex}`}>{piece}</React.Fragment>;
+    });
+  });
+}
+
+function renderTomlLine(line: string, index: number) {
+  const key = `toml-${index}`;
+  if (line.trim().startsWith("#")) {
+    return <span className="toml-comment">{line}</span>;
+  }
+  if (/^\s*\[[^\]]+\]\s*$/.test(line)) {
+    return <span className="toml-section">{line}</span>;
+  }
+  const eqIndex = line.indexOf("=");
+  if (eqIndex > -1) {
+    const left = line.slice(0, eqIndex);
+    const right = line.slice(eqIndex + 1);
+    return (
+      <>
+        <span className="toml-key">{left}</span>
+        <span className="toml-eq">=</span>
+        {renderTomlValue(right, key)}
+      </>
+    );
+  }
+  return <>{line}</>;
+}
+
+function TomlPreview({ text }: { text: string }) {
+  return (
+    <pre className="toml-preview" aria-label="TOML preview">
+      {text.split("\n").map((line, index) => (
+        <div className="toml-line" key={index}>
+          <span className="toml-line-no">{index + 1}</span>
+          <code>{renderTomlLine(line, index)}</code>
+        </div>
+      ))}
+    </pre>
+  );
+}
+
+function fitCodeEditorHeight(editor: HTMLTextAreaElement | null, minHeight: number) {
+  if (!editor) return;
+  editor.style.height = "auto";
+  editor.style.height = `${Math.max(minHeight, editor.scrollHeight + 2)}px`;
+}
+
+function normalizedConfigDirForComparison(value: string) {
+  let normalized = value.trim().replace(/\\/g, "/").replace(/\/+$/, "");
+  if (/^(?:[a-z]:\/|\/\/)/i.test(normalized)) normalized = normalized.toLowerCase();
+  return normalized;
+}
+
+function CodexStateLoading({ lang, loading }: { lang: Lang; loading: boolean }) {
+  return (
+    <section className="cx-state-loading" role="status" aria-live="polite">
+      {loading && <Loader2 className="spin" size={22} aria-hidden="true" />}
+      <span>{loading
+        ? (lang === "zh" ? "正在读取 Codex 配置..." : "Loading Codex configuration...")
+        : (lang === "zh" ? "Codex 配置读取失败，请返回概览重试" : "Could not load the Codex configuration. Retry from Overview.")}</span>
+    </section>
+  );
+}
+
+export function App() {
+  const [lang, setLang] = React.useState<Lang>(getLanguage);
+  const [theme, setTheme] = React.useState<AppTheme>(getTheme);
+  const t = dict[lang];
+  const updater = useAppUpdater();
+  const isMacRuntime = navigator.userAgent.toLowerCase().includes("mac");
+  const [tab, setTab] = React.useState<Tab>("dashboard");
+  const [visitedTabs, setVisitedTabs] = React.useState<Set<Tab>>(() => new Set(["dashboard"]));
+  const [providerMode, setProviderMode] = React.useState<ProviderMode>("list");
+  const [instructionMode, setInstructionMode] = React.useState<InstructionMode>("list");
+  const [promptInjectionMode, setPromptInjectionMode] = React.useState<PromptInjectionMode>(getPromptInjectionMode);
+  const [skillsMcpTab, setSkillsMcpTab] = React.useState<"mcp" | "skills">("mcp");
+  const [editingProviderId, setEditingProviderId] = React.useState<string | null>(null);
+  const [copyingProviderId, setCopyingProviderId] = React.useState<string | null>(null);
+  const [editingOfficialAccountId, setEditingOfficialAccountId] = React.useState<string | null>(null);
+  const [editingDetectedProvider, setEditingDetectedProvider] = React.useState(false);
+  const [editingPromptId, setEditingPromptId] = React.useState<string | null>(null);
+  const [editingBuiltinPrompt, setEditingBuiltinPrompt] = React.useState<BuiltinPromptDetail | null>(null);
+  const [savedProviders, setSavedProviders] = React.useState<SavedProvider[]>([]);
+  const [officialAccounts, setOfficialAccounts] = React.useState<OfficialAccountSummary[]>([]);
+  const [activeProviderId, setActiveProviderId] = React.useState(getActiveProviderId);
+  const [savedPrompts, setSavedPrompts] = React.useState<SavedPrompt[]>([]);
+  const [builtinPromptStatus, setBuiltinPromptStatus] = React.useState<BuiltinPromptStatus[]>([]);
+  const [aboutInfo, setAboutInfo] = React.useState<AboutInfo | null>(null);
+  const [aboutLoading, setAboutLoading] = React.useState(false);
+  const [releaseInfo, setReleaseInfo] = React.useState<ReleaseInfo>({ status: "idle" });
+  const [updatePromptOpen, setUpdatePromptOpen] = React.useState(false);
+  const [sessionStatus, setSessionStatus] = React.useState<SessionSyncStatus | null>(null);
+  const [skillsMcpState, setSkillsMcpState] = React.useState<SkillsMcpState | null>(null);
+  const [skillsMcpImportPreview, setSkillsMcpImportPreview] = React.useState<SkillsMcpImportPreview | null>(null);
+  const [skillsMcpImportOpen, setSkillsMcpImportOpen] = React.useState(false);
+  const [startupDiagnostics, setStartupDiagnostics] = React.useState<StartupDiagnostics | null>(null);
+  const [startupWizardOpen, setStartupWizardOpen] = React.useState(() => !hasSeenStartupWizard());
+  const [startupClosing, setStartupClosing] = React.useState(false);
+  const [sessionQuery, setSessionQuery] = React.useState("");
+  const deferredSessionQuery = React.useDeferredValue(sessionQuery);
+  const [sessionGroupByCwd, setSessionGroupByCwd] = React.useState(false);
+  const [showInternalSessions, setShowInternalSessions] = React.useState(false);
+  const [selectedSessionIds, setSelectedSessionIds] = React.useState<string[]>([]);
+  const [sessionDeleteConfirmOpen, setSessionDeleteConfirmOpen] = React.useState(false);
+  const [sessionDeleteBusy, setSessionDeleteBusy] = React.useState(false);
+  const [sessionDeleteSafetyConfirmed, setSessionDeleteSafetyConfirmed] = React.useState(false);
+  const [state, setState] = React.useState<CodexState | null>(null);
+  const [configDir, setConfigDir] = React.useState("");
+  const [configDirDraft, setConfigDirDraft] = React.useState("");
+  const [loading, setLoading] = React.useState(false);
+  const [refreshing, setRefreshing] = React.useState(true);
+  const [toast, setToast] = React.useState<string>("");
+  const [error, setError] = React.useState<string>("");
+  const [providerForm, setProviderForm] = React.useState<SavedProvider>(defaultProviderForm);
+  const [providerTomlDraft, setProviderTomlDraft] = React.useState("");
+  const [providerTomlDirty, setProviderTomlDirty] = React.useState(false);
+  const [providerDraftRefreshToken, setProviderDraftRefreshToken] = React.useState(0);
+  const [providerApiKeyVisible, setProviderApiKeyVisible] = React.useState(false);
+  const [providerTestingId, setProviderTestingId] = React.useState("");
+  const [availableProviderModels, setAvailableProviderModels] = React.useState<ProviderModel[]>([]);
+  const [providerModelsLoading, setProviderModelsLoading] = React.useState(false);
+  const [actionBusy, setActionBusy] = React.useState<string>("");
+  const [isRestartingCodex, setIsRestartingCodex] = React.useState(false);
+  const [autoRestartCodex, setAutoRestartCodex] = React.useState(
+    getAutoRestartCodex,
+  );
+  const [promptSyncing, setPromptSyncing] = React.useState(false);
+  const [promptCatalogReady, setPromptCatalogReady] = React.useState(false);
+  const [promptForm, setPromptForm] = React.useState<SavedPrompt>(blankPromptForm);
+  const [officialForm, setOfficialForm] = React.useState({
+    name: "",
+    model: "gpt-5.5",
+    authJson: "",
+    configText: buildOfficialTomlPreview("gpt-5.5"),
+  });
+  const [promptModeHelpOpen, setPromptModeHelpOpen] = React.useState(false);
+  const autoUpdateCheckedRef = React.useRef(false);
+  const promptImportRef = React.useRef<HTMLInputElement | null>(null);
+  const skillZipImportRef = React.useRef<HTMLInputElement | null>(null);
+  const officialAuthEditorRef = React.useRef<HTMLTextAreaElement | null>(null);
+  const officialTomlEditorRef = React.useRef<HTMLTextAreaElement | null>(null);
+  const providerTomlEditorRef = React.useRef<HTMLTextAreaElement | null>(null);
+  const providerModelsRequestRef = React.useRef(0);
+  const providerDraftRequestRef = React.useRef(0);
+  const officialDraftRequestRef = React.useRef(0);
+  const loadingGenerationRef = React.useRef(0);
+  const loadingTokensRef = React.useRef(new Set<number>());
+  const actionBusyGenerationRef = React.useRef(0);
+  const actionBusyTokensRef = React.useRef(new Map<number, string>());
+  const promptModeHelpRef = React.useRef<HTMLDivElement | null>(null);
+  const promptRefreshRequestRef = React.useRef(0);
+  const refreshRequestRef = React.useRef(0);
+  const aboutLoadKeyRef = React.useRef("");
+  const sessionAutoLoadKeyRef = React.useRef("");
+  const sessionLoadRequestRef = React.useRef(0);
+  const promptRefreshInFlightRef = React.useRef<Promise<BuiltinPromptStatus[]> | null>(null);
+  const promptAutoRefreshAttemptedRef = React.useRef(false);
+  const promptCatalogReadyRef = React.useRef(false);
+  const promptModeSyncedRef = React.useRef("");
+  const skillsMcpLoadedRef = React.useRef("");
+  const skillsMcpAutoLoadAttemptedRef = React.useRef("");
+  const skillsMcpRequestRef = React.useRef(0);
+  const activeConfigDirKeyRef = React.useRef("");
+  const themeTransitionTimerRef = React.useRef<number | null>(null);
+  const restartingCodexRef = React.useRef(false);
+  const providerTomlPreview = React.useMemo(() => buildProviderTomlPreview(providerForm), [providerForm]);
+  const providerAuthPreview = React.useMemo(() => buildProviderAuthPreview(providerForm), [providerForm]);
+  const activeBuiltinTemplateId = state?.instructionTemplateKey?.startsWith("builtin:")
+    ? state.instructionTemplateKey.slice("builtin:".length)
+    : "";
+  const instructionTemplates = React.useMemo<InstructionTemplate[]>(() => {
+    if (!builtinPromptStatus.length) return bundledInstructionTemplates;
+    return builtinPromptStatus
+      .filter((item) => item.contentSource !== "removed" || item.id === activeBuiltinTemplateId)
+      .map(({ id, filename, title, subtitle, badge }) => ({ id, filename, title, subtitle, badge }));
+  }, [activeBuiltinTemplateId, builtinPromptStatus]);
+  const missingActiveBuiltinTemplateId = activeBuiltinTemplateId
+    && !instructionTemplates.some((item) => item.id === activeBuiltinTemplateId)
+    ? activeBuiltinTemplateId
+    : "";
+
+  const syncActionBusy = React.useCallback(() => {
+    let latestToken = -1;
+    let latestAction = "";
+    for (const [token, action] of actionBusyTokensRef.current) {
+      if (token > latestToken) {
+        latestToken = token;
+        latestAction = action;
+      }
+    }
+    setActionBusy(latestAction);
+  }, []);
+
+  const beginLoading = React.useCallback(() => {
+    const token = ++loadingGenerationRef.current;
+    loadingTokensRef.current.add(token);
+    setLoading(true);
+    return token;
+  }, []);
+
+  const endLoading = React.useCallback((token: number) => {
+    if (!loadingTokensRef.current.delete(token)) return;
+    setLoading(loadingTokensRef.current.size > 0);
+  }, []);
+
+  const beginActionBusy = React.useCallback((action: string) => {
+    const token = ++actionBusyGenerationRef.current;
+    actionBusyTokensRef.current.set(token, action);
+    setActionBusy(action);
+    return token;
+  }, []);
+
+  const endActionBusy = React.useCallback((token: number) => {
+    if (!actionBusyTokensRef.current.delete(token)) return;
+    syncActionBusy();
+  }, [syncActionBusy]);
+
+  const clearActionBusy = React.useCallback((action: string) => {
+    let changed = false;
+    for (const [token, activeAction] of actionBusyTokensRef.current) {
+      if (activeAction !== action) continue;
+      actionBusyTokensRef.current.delete(token);
+      changed = true;
+    }
+    if (changed) syncActionBusy();
+  }, [syncActionBusy]);
+  const currentInstructionId = instructionIdFromPath(state?.instructionFile, instructionTemplates);
+  const releaseStatusLabel = React.useMemo(() => {
+    if (updater.state.phase === "downloading") return lang === "zh" ? "下载中" : "Downloading";
+    if (updater.state.phase === "installing") return lang === "zh" ? "安装中" : "Installing";
+    if (updater.state.phase === "ready") return lang === "zh" ? "等待重启" : "Restart required";
+    if (releaseInfo.status === "checking") return lang === "zh" ? "检查中" : "Checking";
+    if (releaseInfo.status === "error") return lang === "zh" ? "失败" : "Failed";
+    if (releaseInfo.hasUpdate) return lang === "zh" ? "有更新" : "Update found";
+    if (releaseInfo.status === "ok") return lang === "zh" ? "已是最新" : "Up to date";
+    return lang === "zh" ? "未检查" : "Idle";
+  }, [lang, releaseInfo.hasUpdate, releaseInfo.status, updater.state.phase]);
+
+  React.useEffect(() => {
+    persistLanguage(lang);
+  }, [lang]);
+
+  React.useLayoutEffect(() => {
+    document.documentElement.dataset.theme = theme;
+    persistTheme(theme);
+  }, [theme]);
+
+  React.useEffect(() => () => {
+    if (themeTransitionTimerRef.current !== null) {
+      window.clearTimeout(themeTransitionTimerRef.current);
+    }
+    document.documentElement.classList.remove("cx-theme-view-transition", "cx-theme-fallback-transition");
+  }, []);
+
+  const toggleTheme = React.useCallback(() => {
+    const nextTheme: AppTheme = theme === "dark" ? "light" : "dark";
+    const root = document.documentElement;
+    const commitTheme = () => {
+      flushSync(() => setTheme(nextTheme));
+    };
+
+    if (window.matchMedia("(prefers-reduced-motion: reduce)").matches) {
+      commitTheme();
+      return;
+    }
+
+    const transitionDocument = document as ThemeTransitionDocument;
+    if (typeof transitionDocument.startViewTransition === "function") {
+      root.classList.remove("cx-theme-fallback-transition");
+      root.classList.add("cx-theme-view-transition");
+      try {
+        const transition = transitionDocument.startViewTransition(commitTheme);
+        const clearTransitionClass = () => root.classList.remove("cx-theme-view-transition");
+        void transition.finished.then(clearTransitionClass, clearTransitionClass);
+        return;
+      } catch {
+        root.classList.remove("cx-theme-view-transition");
+      }
+    }
+
+    root.classList.add("cx-theme-fallback-transition");
+    commitTheme();
+    if (themeTransitionTimerRef.current !== null) {
+      window.clearTimeout(themeTransitionTimerRef.current);
+    }
+    themeTransitionTimerRef.current = window.setTimeout(() => {
+      root.classList.remove("cx-theme-fallback-transition");
+      themeTransitionTimerRef.current = null;
+    }, 260);
+  }, [theme]);
+
+  React.useEffect(() => {
+    persistPromptInjectionMode(promptInjectionMode);
+  }, [promptInjectionMode]);
+
+  React.useEffect(() => {
+    persistAutoRestartCodex(autoRestartCodex);
+  }, [autoRestartCodex]);
+
+  React.useEffect(() => {
+    activeConfigDirKeyRef.current = normalizedConfigDirForComparison(configDir);
+  }, [configDir]);
+
+  React.useEffect(() => {
+    if (error) setToast("");
+  }, [error]);
+
+  React.useEffect(() => {
+    if (!promptModeHelpOpen) return undefined;
+    const handlePointerDown = (event: PointerEvent) => {
+      const target = event.target;
+      if (target instanceof Node && promptModeHelpRef.current?.contains(target)) return;
+      setPromptModeHelpOpen(false);
+    };
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.key === "Escape") setPromptModeHelpOpen(false);
+    };
+    document.addEventListener("pointerdown", handlePointerDown);
+    document.addEventListener("keydown", handleKeyDown);
+    return () => {
+      document.removeEventListener("pointerdown", handlePointerDown);
+      document.removeEventListener("keydown", handleKeyDown);
+    };
+  }, [promptModeHelpOpen]);
+
+  React.useLayoutEffect(() => {
+    if (providerMode !== "form") return;
+    fitCodeEditorHeight(providerTomlEditorRef.current, 560);
+  }, [providerMode, providerTomlDraft]);
+
+  React.useEffect(() => {
+    if (providerMode !== "form") {
+      providerDraftRequestRef.current += 1;
+      return;
+    }
+    const requestId = ++providerDraftRequestRef.current;
+    if (providerTomlDirty) return;
+
+    const fallback = providerForm.tomlConfig?.trim()
+      || state?.configText?.trim()
+      || providerTomlPreview;
+    if (!providerForm.providerName.trim() || !providerForm.baseUrl.trim() || !providerForm.model.trim()) {
+      setProviderTomlDraft(fallback);
+      return;
+    }
+
+    void providersApi.buildProviderTomlDraft({
+      ...providerForm,
+      id: providerForm.id || customProviderId(providerForm.providerName || providerForm.baseUrl),
+    }, configDir || null)
+      .then((draft) => {
+        if (requestId === providerDraftRequestRef.current) setProviderTomlDraft(draft);
+      })
+      .catch(() => {
+        if (requestId === providerDraftRequestRef.current) setProviderTomlDraft(fallback);
+      });
+  }, [configDir, providerDraftRefreshToken, providerForm, providerMode, providerTomlDirty, providerTomlPreview, state?.configText]);
+
+  React.useEffect(() => {
+    if (tab === "provider" && providerMode === "form") return;
+    providerModelsRequestRef.current += 1;
+    setAvailableProviderModels([]);
+    setProviderModelsLoading(false);
+    if (providerModelsLoading) setToast("");
+  }, [providerMode, providerModelsLoading, tab]);
+
+  React.useEffect(() => {
+    if (tab === "provider" && providerMode === "official") return;
+    officialDraftRequestRef.current += 1;
+    clearActionBusy("loadOfficialDraft");
+  }, [clearActionBusy, providerMode, tab]);
+
+  React.useEffect(() => {
+    if (providerMode !== "official") return;
+    const fit = () => {
+      fitCodeEditorHeight(officialTomlEditorRef.current, 360);
+      fitCodeEditorHeight(officialAuthEditorRef.current, 420);
+    };
+    const frame = window.requestAnimationFrame(fit);
+    window.addEventListener("resize", fit);
+    return () => {
+      window.cancelAnimationFrame(frame);
+      window.removeEventListener("resize", fit);
+    };
+  }, [officialForm.authJson, officialForm.configText, providerMode]);
+
+  React.useEffect(() => {
+    if (!state || promptModeSyncedRef.current === state.codexDir) return;
+    promptModeSyncedRef.current = state.codexDir;
+    if (state.instructionInjectionMode) {
+      setPromptInjectionMode(state.instructionInjectionMode);
+    }
+  }, [state]);
+
+  React.useEffect(() => {
+    setVisitedTabs((tabs) => {
+      if (tabs.has(tab)) return tabs;
+      const next = new Set(tabs);
+      next.add(tab);
+      return next;
+    });
+  }, [tab]);
+
+  const currentProvider = state?.providers.find((p) => p.isCurrent);
+  const liveProviderId = (state?.modelProvider || "openai").trim();
+  const liveProviderApiKey = React.useMemo(() => {
+    const configKey = extractTomlProviderApiKey(state?.configText, liveProviderId);
+    if (!state?.isOfficialProvider) return configKey;
+    return extractOpenAiApiKey(state?.authText).trim();
+  }, [liveProviderId, state?.authText, state?.configText, state?.isOfficialProvider]);
+  const inferredActiveProviderId = React.useMemo(() => {
+    if (state?.isOfficialProvider) return "";
+    if (state?.activeSavedProviderId && savedProviders.some((item) => item.id === state.activeSavedProviderId)) {
+      return state.activeSavedProviderId;
+    }
+    return liveProviderId !== "custom" ? liveProviderId : "";
+  }, [liveProviderId, savedProviders, state?.activeSavedProviderId, state?.isOfficialProvider]);
+  const effectiveActiveProviderId = state?.isOfficialProvider ? "" : inferredActiveProviderId;
+  const currentInstructionPath = (state?.instructionFile || "").replace(/\\/g, "/");
+  const currentInstructionFilename = currentInstructionPath.split("/").pop() || "";
+  const activeInstructionTitle = React.useMemo(() => {
+    const templateKey = state?.instructionTemplateKey || "";
+    if (templateKey.startsWith("builtin:")) {
+      const id = templateKey.slice("builtin:".length);
+      return instructionTemplates.find((item) => item.id === id)?.title || id;
+    }
+    if (templateKey.startsWith("saved:")) {
+      const id = templateKey.slice("saved:".length);
+      return savedPrompts.find((item) => item.id === id)?.title || id;
+    }
+    return savedPrompts.find((item) => item.filename === currentInstructionFilename)?.title
+      || instructionTemplates.find((item) => item.filename === currentInstructionFilename)?.title
+      || currentInstructionFilename
+      || (lang === "zh" ? "当前提示词" : "Current prompt");
+  }, [currentInstructionFilename, instructionTemplates, lang, savedPrompts, state?.instructionTemplateKey]);
+  const canonicalSavedProviders = React.useMemo(() => {
+    const groups = new Map<string, SavedProvider[]>();
+    savedProviders.forEach((provider) => {
+      const identity = providerIdentityKey(provider, savedProviderApiKey(provider));
+      const key = identity || `id:${provider.id}`;
+      const group = groups.get(key);
+      if (group) group.push(provider);
+      else groups.set(key, [provider]);
+    });
+    return Array.from(groups.values()).map((group) =>
+      group.find((item) => item.id === effectiveActiveProviderId)
+      || group.find((item) => item.id === activeProviderId)
+      || group[0],
+    );
+  }, [activeProviderId, effectiveActiveProviderId, savedProviders]);
+
+  const detectedRows = React.useMemo(() => {
+    if (state?.isOfficialProvider) return [];
+    return (state?.providers || []).filter((p) => p.isCurrent).map((p) => {
+      const configKey = extractTomlProviderApiKey(state?.configText, p.id);
+      const apiKey = p.isCurrent ? liveProviderApiKey || configKey : configKey;
+      return {
+        id: `detected-${p.id}`,
+        source: "detected" as const,
+        providerName: p.name || p.id,
+        baseUrl: p.baseUrl || "",
+        model: state?.model || "gpt-5.5",
+        apiKey,
+        wireApi: p.wireApi || "responses",
+        requiresOpenaiAuth: p.requiresOpenaiAuth ?? false,
+        isCurrent: p.isCurrent,
+      };
+    });
+  }, [liveProviderApiKey, state?.configText, state?.isOfficialProvider, state?.model, state?.providers]);
+
+  const localRows = React.useMemo(() => {
+    return canonicalSavedProviders.map((p) => ({
+      ...p,
+      source: "local" as const,
+      isCurrent: effectiveActiveProviderId === p.id,
+    }));
+  }, [canonicalSavedProviders, effectiveActiveProviderId]);
+
+  const providerPageRows = React.useMemo<ProviderRow[]>(() => {
+    const selectedOfficial = officialAccounts.find((account) => account.selected);
+    const officialRows: ProviderRow[] = officialAccounts.length === 0
+      ? [{
+        id: "openai-official",
+        source: "official",
+        providerName: "OpenAI Official",
+        baseUrl: state?.model || "official",
+        model: state?.model || "official",
+        wireApi: "official",
+        requiresOpenaiAuth: false,
+        isCurrent: Boolean(state?.isOfficialProvider),
+        sourceLabel: lang === "zh" ? "兼容模式" : "Legacy mode",
+        editable: true,
+        deletable: false,
+        testable: false,
+        officialKind: "legacy",
+      }]
+      : officialAccounts.map((account) => ({
+        id: `openai-official-${account.id}`,
+        source: "official",
+        providerName: `OpenAI Official · ${account.name}`,
+        baseUrl: account.model || "official",
+        model: account.model || "official",
+        wireApi: "official",
+        requiresOpenaiAuth: false,
+        isCurrent: Boolean(state?.isOfficialProvider && account.selected),
+        sourceLabel: lang === "zh" ? "已保存账号" : "Saved account",
+        editable: true,
+        deletable: true,
+        testable: false,
+        officialAccountId: account.id,
+        officialKind: "saved",
+      }));
+    if (officialAccounts.length > 0 && state?.isOfficialProvider && !selectedOfficial) {
+      officialRows.unshift({
+        id: "openai-official-unsaved",
+        source: "official",
+        providerName: lang === "zh" ? "OpenAI Official · 未保存登录" : "OpenAI Official · Unsaved login",
+        baseUrl: state.model || "official",
+        model: state.model || "official",
+        wireApi: "official",
+        requiresOpenaiAuth: false,
+        isCurrent: true,
+        sourceLabel: lang === "zh" ? "未保存" : "Unsaved",
+        editable: false,
+        deletable: false,
+        testable: false,
+        officialKind: "unsaved",
+      });
+    }
+
+    const thirdPartyRows: ProviderRow[] = [];
+    const seen = new Set<string>();
+    localRows.forEach((row) => {
+      const key = providerIdentityKey(row, savedProviderApiKey(row));
+      if (key) seen.add(key);
+      thirdPartyRows.push({
+        ...row,
+        editable: true,
+        deletable: true,
+        testable: true,
+        testingKey: `${row.source}-${row.id}`,
+      });
+    });
+    detectedRows.forEach((row) => {
+      if (row.id === "detected-custom" && inferredActiveProviderId) return;
+      const key = providerIdentityKey(row, row.apiKey);
+      if (key && seen.has(key)) return;
+      if (key) seen.add(key);
+      thirdPartyRows.push({
+        ...row,
+        editable: true,
+        deletable: false,
+        testable: true,
+        testingKey: `${row.source}-${row.id}`,
+      });
+    });
+    return [...officialRows, ...thirdPartyRows];
+  }, [detectedRows, inferredActiveProviderId, lang, localRows, officialAccounts, state?.isOfficialProvider, state?.model]);
+
+  const findLocalProviderForRow = React.useCallback((row: ProviderRow) => {
+    if (row.source === "official") return undefined;
+    return canonicalSavedProviders.find((item) =>
+      row.source === "local"
+        ? item.id === row.id
+        : providerIdentityKey(item, savedProviderApiKey(item))
+          === providerIdentityKey(row, row.apiKey),
+    );
+  }, [canonicalSavedProviders]);
+
+  const visibleSessions = React.useMemo(
+    () => (sessionStatus?.sessions || []).filter((item) => showInternalSessions || !item.isSubagent),
+    [sessionStatus?.sessions, showInternalSessions],
+  );
+
+  const filteredSessions = React.useMemo(() => {
+    const query = deferredSessionQuery.trim().toLowerCase();
+    if (!query) return visibleSessions;
+    return visibleSessions.filter((item) => [item.title, item.cwd, item.rolloutPath, item.modelProvider, item.model, item.id]
+      .filter(Boolean)
+      .some((value) => String(value).toLowerCase().includes(query)));
+  }, [deferredSessionQuery, visibleSessions]);
+
+  const allSessionsByCwd = React.useMemo(() => {
+    const groups = new Map<string, SessionPreview[]>();
+    for (const item of visibleSessions) {
+      const key = item.cwd || (lang === "zh" ? "未记录工作目录" : "No workspace recorded");
+      if (!groups.has(key)) groups.set(key, []);
+      groups.get(key)!.push(item);
+    }
+    return groups;
+  }, [lang, visibleSessions]);
+
+  const groupedSessions = React.useMemo(() => {
+    const groups = new Map<string, SessionPreview[]>();
+    if (!sessionGroupByCwd) {
+      groups.set(lang === "zh" ? "全部会话" : "All sessions", filteredSessions);
+      return Array.from(groups.entries());
+    }
+    for (const item of filteredSessions) {
+      const key = item.cwd || (lang === "zh" ? "未记录工作目录" : "No workspace recorded");
+      if (!groups.has(key)) groups.set(key, []);
+      groups.get(key)!.push(item);
+    }
+    return Array.from(groups.entries()).sort((a, b) => b[1].length - a[1].length);
+  }, [filteredSessions, lang, sessionGroupByCwd]);
+
+  const sessionHasMismatches = Boolean(sessionStatus?.needsSync);
+  const sessionTargetLabel = lang === "zh" ? "共享会话" : "Shared history";
+  const sessionSyncCount = sessionMismatchCount(sessionStatus);
+  const sessionVisibleTotal = showInternalSessions
+    ? (sessionStatus?.topLevelThreads ?? 0) + (sessionStatus?.subagentThreads ?? 0)
+    : (sessionStatus?.topLevelThreads ?? 0);
+  const sessionPreviewTruncated = sessionVisibleTotal > visibleSessions.length;
+  const selectedSessionSet = React.useMemo(() => new Set(selectedSessionIds), [selectedSessionIds]);
+  const selectedSessions = React.useMemo(
+    () => (sessionStatus?.sessions || []).filter((item) => selectedSessionSet.has(item.id)),
+    [selectedSessionSet, sessionStatus?.sessions],
+  );
+
+  React.useEffect(() => {
+    setSelectedSessionIds((ids) => ids.filter((id) => (sessionStatus?.sessions || []).some((item) => item.id === id)));
+  }, [sessionStatus?.sessions]);
+
+  React.useEffect(() => {
+    if (sessionDeleteConfirmOpen && selectedSessions.length === 0) {
+      setSessionDeleteConfirmOpen(false);
+    }
+  }, [selectedSessions.length, sessionDeleteConfirmOpen]);
+
+  const call = React.useCallback(async <T,>(fn: () => Promise<T>, success?: (data: T) => void) => {
+    const loadingToken = beginLoading();
+    setError("");
+    try {
+      const data = await fn();
+      success?.(data);
+    } catch (e) {
+      setError(String(e));
+    } finally {
+      endLoading(loadingToken);
+    }
+  }, [beginLoading, endLoading]);
+
+  const refresh = React.useCallback((includeDiagnostics: boolean) => {
+    const requestId = ++refreshRequestRef.current;
+    officialDraftRequestRef.current += 1;
+    clearActionBusy("loadOfficialDraft");
+    skillsMcpRequestRef.current += 1;
+    skillsMcpLoadedRef.current = "";
+    skillsMcpAutoLoadAttemptedRef.current = "";
+    const requestedConfigDir = configDirDraft.trim();
+    const resolvedConfigDir = requestedConfigDir || null;
+    const activeCodexDir = state?.codexDir || configDir;
+    setRefreshing(true);
+    setError("");
+    setState(null);
+    setSessionStatus(null);
+    setSkillsMcpState(null);
+    setSkillsMcpImportOpen(false);
+    setSkillsMcpImportPreview(null);
+    setAboutInfo(null);
+    aboutLoadKeyRef.current = "";
+    sessionAutoLoadKeyRef.current = "";
+    sessionLoadRequestRef.current += 1;
+
+    if (includeDiagnostics) {
+      void codexApi.getStartupDiagnostics(resolvedConfigDir)
+        .then((diagnostics) => {
+          if (requestId === refreshRequestRef.current) setStartupDiagnostics(diagnostics);
+        })
+        .catch(() => {
+          if (requestId === refreshRequestRef.current) setStartupDiagnostics(null);
+        });
+    }
+
+    void codexApi.getCodexState(resolvedConfigDir)
+      .then((next) => {
+        if (requestId !== refreshRequestRef.current) return;
+        if (normalizedConfigDirForComparison(next.codexDir)
+          !== normalizedConfigDirForComparison(activeCodexDir)) {
+          providerDraftRequestRef.current += 1;
+          providerModelsRequestRef.current += 1;
+          setProviderMode("list");
+          setEditingProviderId(null);
+          setCopyingProviderId(null);
+          setEditingOfficialAccountId(null);
+          setEditingDetectedProvider(false);
+          setProviderTomlDirty(false);
+          setProviderTomlDraft("");
+          setAvailableProviderModels([]);
+          setProviderModelsLoading(false);
+        }
+        activeConfigDirKeyRef.current = normalizedConfigDirForComparison(next.codexDir);
+        setConfigDir(next.codexDir);
+        setConfigDirDraft(next.codexDir);
+        setState(next);
+        setRefreshing(false);
+
+        void Promise.allSettled([
+          providersApi.listSavedProviders(),
+          officialAccountsApi.listOfficialAccounts(next.codexDir),
+          promptsApi.listSavedPrompts(),
+          promptsApi.getBuiltinPromptStatus(),
+        ]).then(([providers, accounts, prompts, promptStatus]) => {
+          if (requestId !== refreshRequestRef.current) return;
+          if (providers.status === "fulfilled") setSavedProviders(providers.value);
+          if (accounts.status === "fulfilled") setOfficialAccounts(accounts.value);
+          else setError(String(accounts.reason));
+          if (prompts.status === "fulfilled") setSavedPrompts(prompts.value);
+          if (promptStatus.status === "fulfilled") {
+            setBuiltinPromptStatus(uniqueBuiltinPromptStatuses(promptStatus.value));
+          }
+        });
+      })
+      .catch((nextError) => {
+        if (requestId !== refreshRequestRef.current) return;
+        setRefreshing(false);
+        setError(String(nextError));
+      });
+  }, [clearActionBusy, configDir, configDirDraft, state?.codexDir]);
+
+  React.useEffect(() => {
+    refresh(startupWizardOpen);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  React.useEffect(() => {
+    if (!state?.codexDir) return;
+    const loadKey = state.codexDir;
+    if (aboutLoadKeyRef.current === loadKey) return;
+    aboutLoadKeyRef.current = loadKey;
+    setAboutLoading(true);
+    void codexApi.getAboutInfo(loadKey)
+      .then((about) => {
+        if (aboutLoadKeyRef.current === loadKey) setAboutInfo(about);
+      })
+      .catch((aboutError) => {
+        if (aboutLoadKeyRef.current !== loadKey) return;
+        aboutLoadKeyRef.current = "";
+        setError(String(aboutError));
+      })
+      .finally(() => {
+        if (aboutLoadKeyRef.current === loadKey || aboutLoadKeyRef.current === "") {
+          setAboutLoading(false);
+        }
+      });
+  }, [state?.codexDir, tab]);
+
+  React.useEffect(() => {
+    if (!state) return;
+    if (state.isOfficialProvider) {
+      if (activeProviderId) {
+        persistActiveProviderId("");
+        setActiveProviderId("");
+      }
+      return;
+    }
+    if (!savedProviders.length) return;
+    const mappedProviderId = savedProviders.some((item) => item.id === inferredActiveProviderId)
+      ? inferredActiveProviderId
+      : "";
+    if (mappedProviderId && mappedProviderId !== activeProviderId) {
+      persistActiveProviderId(mappedProviderId);
+      setActiveProviderId(mappedProviderId);
+      return;
+    }
+    if (activeProviderId && !savedProviders.some((item) => item.id === activeProviderId)) {
+      persistActiveProviderId("");
+      setActiveProviderId("");
+    }
+  }, [activeProviderId, inferredActiveProviderId, liveProviderId, savedProviders, state]);
+
+  const handleActionResult = (result: ActionResult) => {
+    setState(result.state);
+    setSessionStatus(null);
+    sessionAutoLoadKeyRef.current = "";
+    sessionLoadRequestRef.current += 1;
+    setToast(result.message);
+    void Promise.allSettled([
+      promptsApi.listSavedPrompts(),
+      providersApi.listSavedProviders(),
+      officialAccountsApi.listOfficialAccounts(result.state.codexDir),
+    ])
+      .then(([prompts, providers, accounts]) => {
+        if (prompts.status === "fulfilled") setSavedPrompts(prompts.value);
+        if (providers.status === "fulfilled") setSavedProviders(providers.value);
+        if (accounts.status === "fulfilled") setOfficialAccounts(accounts.value);
+      })
+      .catch(() => undefined);
+  };
+
+  const restartCodex = React.useCallback(async ({
+    automatic = false,
+    configMessage = "",
+  }: { automatic?: boolean; configMessage?: string } = {}) => {
+    if (restartingCodexRef.current) return;
+    restartingCodexRef.current = true;
+    setIsRestartingCodex(true);
+    setError("");
+    try {
+      const result = await codexApi.restartCodexApp();
+      if (result.success) {
+        setToast(result.message);
+      } else if (automatic) {
+        setError(lang === "zh"
+          ? `${configMessage || "配置已应用"}，但 Codex 自动重启失败。你可以点击“重启 Codex”再次尝试。`
+          : "The configuration was applied, but Codex could not restart automatically. Use Restart Codex to try again.");
+      } else {
+        setError(lang === "zh"
+          ? `Codex 重启失败：${result.message}`
+          : "Codex restart failed. No safely restartable Codex Desktop was found.");
+      }
+    } catch {
+      if (automatic) {
+        setError(lang === "zh"
+          ? `${configMessage || "配置已应用"}，但 Codex 自动重启失败。你可以点击“重启 Codex”再次尝试。`
+          : "The configuration was applied, but Codex could not restart automatically. Use Restart Codex to try again.");
+      } else {
+        setError(lang === "zh" ? "Codex 重启失败，请稍后重试" : "Codex restart failed. Try again later.");
+      }
+    } finally {
+      restartingCodexRef.current = false;
+      setIsRestartingCodex(false);
+    }
+  }, [lang]);
+
+  const handleConfigActionResult = (
+    before: CodexState | null,
+    result: ActionResult,
+  ) => {
+    handleActionResult(result);
+    const liveChanged = Boolean(before)
+      && (before?.configText !== result.state.configText || before?.authText !== result.state.authText);
+    if (autoRestartCodex && liveChanged) {
+      void restartCodex({ automatic: true, configMessage: result.message });
+    }
+  };
+
+  const switchInstructionTemplate = (templateId: string) =>
+    call(
+      () => promptsApi.enableInstructionTemplate(configDir || null, templateId, promptInjectionMode),
+      handleActionResult,
+    );
+
+  const disableInstruction = () =>
+    call(
+      () => promptsApi.disableInstruction(configDir || null),
+      handleActionResult,
+    );
+
+  const disableExternalInstruction = () =>
+    call(
+      () => promptsApi.disableExternalInstruction(configDir || null),
+      handleActionResult,
+    );
+
+  const openAddPrompt = () => {
+    setEditingPromptId(null);
+    setEditingBuiltinPrompt(null);
+    setPromptForm({ ...blankPromptForm });
+    setInstructionMode("form");
+  };
+
+  const openEditPrompt = (prompt: SavedPrompt) => {
+    setEditingPromptId(prompt.id);
+    setEditingBuiltinPrompt(null);
+    setPromptForm(prompt);
+    setInstructionMode("form");
+  };
+
+  const openEditBuiltinPrompt = (templateId: string) =>
+    call(
+      () => promptsApi.getBuiltinPromptDetail(templateId),
+      (detail) => {
+        setEditingPromptId(null);
+        setEditingBuiltinPrompt(detail);
+        setPromptForm({
+          id: detail.id,
+          title: detail.title,
+          filename: detail.filename,
+          content: detail.content,
+        });
+        setInstructionMode("form");
+      },
+    );
+
+  const normalizedPromptForm = (): SavedPrompt => {
+    const existing = savedPrompts.filter((item) => item.id !== editingPromptId);
+    const requestedFilename = promptForm.filename.trim() || `${providerId(promptForm.title || "prompt")}.md`;
+    const filename = editingPromptId ? requestedFilename : uniquePromptFilename(requestedFilename, existing.map((item) => item.filename));
+    return {
+      ...promptForm,
+      id: editingPromptId || uniqueId(promptForm.id || promptForm.title || filename, existing.map((item) => item.id)),
+      title: promptForm.title.trim(),
+      filename,
+      content: promptForm.content,
+    };
+  };
+
+  const savePromptOnly = () => {
+    if (editingBuiltinPrompt) {
+      call(
+        async () => {
+          const detail = await promptsApi.saveBuiltinPromptOverride(
+            editingBuiltinPrompt.id,
+            promptForm.content,
+          );
+          const statuses = await promptsApi.getBuiltinPromptStatus();
+          return { detail, statuses };
+        },
+        ({ detail, statuses }) => {
+          setEditingBuiltinPrompt(detail);
+          setBuiltinPromptStatus(uniqueBuiltinPromptStatuses(statuses));
+          setInstructionMode("list");
+          setToast(detail.customized
+            ? (lang === "zh"
+              ? "本地修改已保存，下次启用时生效；后续 GitHub 同步将跳过这个模板"
+              : "Local changes saved for the next activation. Future GitHub syncs will skip this template.")
+            : (lang === "zh" ? "内容没有变化，模板将继续参与 GitHub 同步" : "No changes detected. This template will continue to sync from GitHub."));
+        },
+      );
+      return;
+    }
+    call(
+      async () => {
+        await promptsApi.savePrompt(normalizedPromptForm());
+        return promptsApi.listSavedPrompts();
+      },
+      (promptList) => {
+        setSavedPrompts(promptList);
+        setInstructionMode("list");
+        setEditingPromptId(null);
+        setToast(lang === "zh" ? "提示词已保存" : "Prompt saved");
+      },
+    );
+  };
+
+  const enableSavedPrompt = (id: string) =>
+    call(() => promptsApi.enableSavedPrompt(configDir || null, id, promptInjectionMode), handleActionResult);
+
+  const removeSavedPrompt = (id: string) =>
+    call(
+      async () => {
+        await promptsApi.deleteSavedPrompt(id);
+        return promptsApi.listSavedPrompts();
+      },
+      (promptList) => {
+        setSavedPrompts(promptList);
+        setToast(lang === "zh" ? "提示词已删除" : "Prompt deleted");
+      },
+    );
+
+  const importPromptMd = async (file?: File | null) => {
+    if (!file) return;
+    if (!file.name.toLowerCase().endsWith(".md")) {
+      setError(lang === "zh" ? "请选择 .md 提示词文件" : "Please choose a .md prompt file");
+      return;
+    }
+    const actionToken = beginActionBusy("importPrompt");
+    const loadingToken = beginLoading();
+    setError("");
+    try {
+      const content = await file.text();
+      const title = file.name.replace(/\.md$/i, "");
+      const filename = uniquePromptFilename(file.name, savedPrompts.map((item) => item.filename));
+      await promptsApi.savePrompt({
+        id: uniqueId(title, savedPrompts.map((item) => item.id)),
+        title: filename.replace(/\.md$/i, ""),
+        filename,
+        content,
+      });
+      const promptList = await promptsApi.listSavedPrompts();
+      setSavedPrompts(promptList);
+      setToast(lang === "zh" ? `已导入提示词：${file.name}` : `Prompt imported: ${file.name}`);
+    } catch (e) {
+      setError(String(e));
+    } finally {
+      endLoading(loadingToken);
+      endActionBusy(actionToken);
+      if (promptImportRef.current) promptImportRef.current.value = "";
+    }
+  };
+
+  const refreshBuiltinPrompts = async ({ quiet = false }: { quiet?: boolean } = {}) => {
+    const requestId = ++promptRefreshRequestRef.current;
+    if (!quiet) promptAutoRefreshAttemptedRef.current = true;
+    if (!quiet) setError("");
+    try {
+      const existingRequest = promptRefreshInFlightRef.current;
+      const request = existingRequest || promptsApi.refreshBuiltinPrompts(configDir || null);
+      if (!existingRequest) {
+        promptRefreshInFlightRef.current = request;
+        setPromptSyncing(true);
+        const clearRequest = () => {
+          if (promptRefreshInFlightRef.current !== request) return;
+          promptRefreshInFlightRef.current = null;
+          setPromptSyncing(false);
+        };
+        void request.then(clearRequest, clearRequest);
+      }
+      const list = await request;
+      if (requestId !== promptRefreshRequestRef.current) return;
+      const uniqueList = uniqueBuiltinPromptStatuses(list);
+      const catalogFailed = uniqueList.some((item) => item.syncIssue === "catalog");
+      const contentFetchFailures = uniqueList.filter((item) =>
+        item.contentSource === "unavailable" || item.syncIssue === "content",
+      ).length;
+      if (!catalogFailed) {
+        promptCatalogReadyRef.current = true;
+        setPromptCatalogReady(true);
+        setBuiltinPromptStatus(uniqueList);
+      } else if (!promptCatalogReadyRef.current) {
+        setBuiltinPromptStatus(uniqueList);
+      }
+      const updated = uniqueList.filter((item) => item.updated).length;
+      if (!quiet) {
+        setToast(catalogFailed
+          ? promptCatalogReadyRef.current
+            ? (lang === "zh" ? "在线模板库暂时不可用，已保留当前列表" : "Online templates are unavailable; keeping the current list")
+            : (lang === "zh" ? "在线模板库暂时不可用，已使用本地模板" : "Online templates are unavailable; using local templates")
+          : contentFetchFailures > 0
+            ? (lang === "zh" ? `模板目录已同步，${contentFetchFailures} 个模板暂用本地内容` : `Template catalog synced; ${contentFetchFailures} template(s) are using local content`)
+          : updated > 0
+            ? (lang === "zh" ? `已同步 ${updated} 个提示词模板` : `${updated} prompt template(s) synced`)
+            : (lang === "zh" ? "提示词模板已是最新" : "Prompt templates are up to date"));
+      }
+    } catch (e) {
+      if (requestId === promptRefreshRequestRef.current) {
+        if (!quiet) setError(String(e));
+      }
+    }
+  };
+
+  const normalizedProviderForm = (tomlConfig = providerTomlDraft || providerForm.tomlConfig || buildProviderTomlPreview(providerForm)): SavedProvider => ({
+    ...providerForm,
+    id: editingProviderId
+      || (copyingProviderId
+        ? providerForm.id
+        : uniqueId(providerForm.id || customProviderId(providerForm.providerName || providerForm.baseUrl), savedProviders.map((item) => item.id))),
+    providerName: providerForm.providerName.trim(),
+    baseUrl: providerForm.baseUrl.trim().replace(/\/+$/, ""),
+    model: providerForm.model.trim(),
+    apiKey: (providerForm.apiKey || "").trim(),
+    tomlConfig: tomlConfig.trimEnd(),
+    wireApi: providerForm.wireApi || "responses",
+    requiresOpenaiAuth: providerForm.requiresOpenaiAuth,
+  });
+
+  const applyProviderConfig = (provider: SavedProvider) => {
+    const tomlConfig = provider.tomlConfig?.trim();
+    if (tomlConfig) {
+      return providersApi.saveProviderTomlConfig(
+        configDir || null,
+        tomlConfig,
+        provider.apiKey || "",
+      );
+    }
+    return providersApi.switchProvider({
+      configDir: configDir || null,
+      providerId: provider.id,
+      providerName: provider.providerName,
+      baseUrl: provider.baseUrl,
+      model: provider.model,
+      apiKey: provider.apiKey || "",
+      wireApi: provider.wireApi,
+      requiresOpenaiAuth: provider.requiresOpenaiAuth,
+    });
+  };
+
+  const saveProviderOnly = () => {
+    const pendingProvider = normalizedProviderForm();
+    if (!pendingProvider.providerName || !pendingProvider.baseUrl || !pendingProvider.model) {
+      setError(lang === "zh"
+        ? "请填写供应商名称、API 请求地址和模型"
+        : "Provider name, API URL, and model are required");
+      return;
+    }
+    const before = state;
+    return call(
+      async () => {
+        let provider = pendingProvider;
+        if (!providerTomlDirty) {
+          const draftSource = providerForm.tomlConfig?.trim() || state?.configText?.trim() || "";
+          const providerForDraft = normalizedProviderForm(draftSource);
+          const latestDraft = await providersApi.buildProviderTomlDraft(
+            providerForDraft,
+            configDir || null,
+          );
+          provider = { ...providerForDraft, tomlConfig: latestDraft.trimEnd() };
+          setProviderTomlDraft(provider.tomlConfig || "");
+        }
+        const applyAfterSave = editingDetectedProvider
+          || Boolean(editingProviderId && editingProviderId === effectiveActiveProviderId);
+        const applied = applyAfterSave
+          ? await providersApi.saveActiveProvider(provider, configDir || null)
+          : null;
+        if (!applyAfterSave) await providersApi.saveProvider(provider);
+        const providerList = await providersApi.listSavedProviders();
+        return { applied, providerList };
+      },
+      ({ applied, providerList }) => {
+        if (applied) handleConfigActionResult(before, applied);
+        setSavedProviders(providerList);
+        setProviderMode("list");
+        setEditingProviderId(null);
+        setCopyingProviderId(null);
+        setEditingDetectedProvider(false);
+        setProviderTomlDirty(false);
+        setToast(applied
+          ? (lang === "zh" ? "供应商配置已保存并热更新" : "Provider saved and hot-applied")
+          : (lang === "zh" ? "供应商配置已保存" : "Provider saved"));
+      },
+    );
+  };
+
+  const switchProvider = (provider: SavedProvider) => {
+    const before = state;
+    return call(
+      () => applyProviderConfig(provider),
+      (result) => {
+        persistActiveProviderId(provider.id);
+        setActiveProviderId(provider.id);
+        handleConfigActionResult(before, result);
+      },
+    );
+  };
+
+  const resetAvailableProviderModels = () => {
+    providerModelsRequestRef.current += 1;
+    setAvailableProviderModels([]);
+    setProviderModelsLoading(false);
+  };
+
+  const fetchProviderModels = async () => {
+    const baseUrl = providerForm.baseUrl.trim();
+    const apiKey = (providerForm.apiKey || "").trim();
+    if (!baseUrl || !apiKey) {
+      setError("");
+      setToast(lang === "zh" ? "请先填写 API 请求地址和 API Key" : "Enter the API URL and API key first");
+      return;
+    }
+
+    const requestId = providerModelsRequestRef.current + 1;
+    providerModelsRequestRef.current = requestId;
+    setProviderModelsLoading(true);
+    setError("");
+    setToast(lang === "zh" ? "正在获取模型列表..." : "Fetching model list...");
+    try {
+      const result = await providersApi.fetchProviderModels(baseUrl, apiKey);
+      if (providerModelsRequestRef.current !== requestId) return;
+      setAvailableProviderModels(result.models);
+      setToast(result.models.length > 0
+        ? (lang === "zh" ? `已获取 ${result.models.length} 个模型` : `${result.models.length} models fetched`)
+        : (lang === "zh" ? "连接成功，但供应商没有返回模型" : "Connected, but the provider returned no models"));
+    } catch (e) {
+      if (providerModelsRequestRef.current !== requestId) return;
+      setToast("");
+      setError(String(e));
+    } finally {
+      if (providerModelsRequestRef.current === requestId) setProviderModelsLoading(false);
+    }
+  };
+
+  const testProvider = async (id: string, baseUrl: string, apiKey?: string | null) => {
+    const actionToken = beginActionBusy("testProvider");
+    setProviderTestingId(id);
+    setError("");
+    setToast(lang === "zh" ? "正在检测连接..." : "Testing connection...");
+    try {
+      const result = await providersApi.testProviderConnection(baseUrl, apiKey || null);
+      if (result.ok) {
+        setToast(lang === "zh" ? `连接成功，响应延迟 ${result.durationMs}ms` : `Connected, ${result.durationMs}ms latency`);
+      } else {
+        setToast("");
+        setError(lang === "zh" ? `连接失败：${result.message}` : `Connection failed: ${result.message}`);
+      }
+    } catch (e) {
+      setToast("");
+      setError(String(e));
+    } finally {
+      setProviderTestingId("");
+      endActionBusy(actionToken);
+    }
+  };
+
+  const saveProviderConfig = saveProviderOnly;
+
+  const switchOfficialProvider = () => {
+    const before = state;
+    return call(
+      () => officialAccountsApi.switchOfficialProvider(configDir || null),
+      (result) => {
+        persistActiveProviderId("");
+        setActiveProviderId("");
+        handleConfigActionResult(before, result);
+      },
+    );
+  };
+
+  const switchOfficialAccount = (accountId: string) => {
+    const before = state;
+    return call(
+      () => officialAccountsApi.switchOfficialAccount(configDir || null, accountId),
+      (result) => {
+        persistActiveProviderId("");
+        setActiveProviderId("");
+        handleConfigActionResult(before, result);
+      },
+    );
+  };
+
+  const prepareNewOfficialAccount = () => {
+    const before = state;
+    return call(
+      () => officialAccountsApi.prepareNewOfficialAccount(configDir || null),
+      (result) => {
+        persistActiveProviderId("");
+        setActiveProviderId("");
+        setEditingOfficialAccountId(null);
+        setProviderMode("list");
+        handleConfigActionResult(before, result);
+      },
+    );
+  };
+
+  const captureCurrentOfficialAccount = async (name: string) => {
+    const loadingToken = beginLoading();
+    setError("");
+    try {
+      const result = await officialAccountsApi.captureCurrentOfficialAccount(configDir || null, name);
+      handleActionResult(result);
+      return true;
+    } catch (captureError) {
+      setError(String(captureError));
+      return false;
+    } finally {
+      endLoading(loadingToken);
+    }
+  };
+
+  const deleteOfficialAccount = async (accountId: string) => {
+    const before = state;
+    const loadingToken = beginLoading();
+    setError("");
+    try {
+      const result = await officialAccountsApi.deleteOfficialAccount(configDir || null, accountId);
+      handleConfigActionResult(before, result);
+      return true;
+    } catch (deleteError) {
+      setError(String(deleteError));
+      return false;
+    } finally {
+      endLoading(loadingToken);
+    }
+  };
+
+  const restoreOfficialProvider = () =>
+    call(
+      async () => {
+        const result = await officialAccountsApi.restoreOfficialProvider(configDir || null);
+        const draft = await officialAccountsApi.getOfficialConfigDraft(configDir || null);
+        return { result, draft };
+      },
+      ({ result, draft }) => {
+        if (draft) {
+          setOfficialForm({
+            name: "",
+            model: draft.model || "gpt-5.5",
+            authJson: draft.authJson,
+            configText: draft.configText || buildOfficialTomlPreview(draft.model || "gpt-5.5"),
+          });
+        }
+        handleActionResult(result);
+        setToast(lang === "zh" ? "已读取保存的官方配置" : "Saved official config loaded");
+      },
+    );
+
+  const loadCcSwitchOfficial = async () => {
+    const requestId = ++officialDraftRequestRef.current;
+    const actionToken = beginActionBusy("loadCcSwitchOfficial");
+    setError("");
+    try {
+      const candidate = await officialAccountsApi.readCcSwitchOfficialAuth();
+      if (requestId !== officialDraftRequestRef.current) return;
+      if (!candidate) {
+        setToast(lang === "zh" ? "未找到 CC Switch 官方配置" : "No CC Switch official config found");
+        return;
+      }
+      setOfficialForm((current) => ({
+        ...current,
+        model: candidate.model || current.model || state?.model || "gpt-5.5",
+        authJson: candidate.authJson,
+        configText: candidate.configText || current.configText,
+      }));
+      setToast(lang === "zh" ? "已从 CC Switch 载入" : "Loaded from CC Switch");
+    } catch (e) {
+      if (requestId === officialDraftRequestRef.current) setError(String(e));
+    } finally {
+      endActionBusy(actionToken);
+    }
+  };
+
+  const resetOfficialProvider = () => {
+    const before = state;
+    return call(
+      () => officialAccountsApi.resetOfficialProvider({
+        configDir: configDir || null,
+        model: officialForm.model,
+        authJson: null,
+        configText: officialForm.configText,
+      }),
+      (result) => {
+        persistActiveProviderId("");
+        setActiveProviderId("");
+        setOfficialForm({
+          name: "",
+          model: result.state.model || officialForm.model || "gpt-5.5",
+          authJson: officialAuthPlaceholder,
+          configText: result.state.configText
+            || officialForm.configText
+            || buildOfficialTomlPreview(result.state.model || officialForm.model || "gpt-5.5"),
+        });
+        handleConfigActionResult(before, result);
+      },
+    );
+  };
+
+  const importFromCcSwitch = async () => {
+    const actionToken = beginActionBusy("importCcSwitch");
+    const loadingToken = beginLoading();
+    setError("");
+    try {
+      const result = await providersApi.importCcSwitchProviders();
+      const warningText = result.skipped > 0
+        ? (lang === "zh" ? `，跳过 ${result.skipped}` : `, ${result.skipped} skipped`)
+        : "";
+      const successText = lang === "zh"
+        ? `cc-switch 导入完成：新增 ${result.added}，更新 ${result.updated}，合并 ${result.merged}${warningText}；未切换当前供应商`
+        : `cc-switch import complete: ${result.added} added, ${result.updated} updated, ${result.merged} merged${warningText}; current provider unchanged`;
+      try {
+        const nextState = await codexApi.getCodexState(configDir || null);
+        setSavedProviders(result.providers);
+        setState(nextState);
+        setToast(successText);
+      } catch (refreshError) {
+        setSavedProviders(result.providers);
+        setToast(lang === "zh"
+          ? `${successText}；状态刷新失败，请手动刷新：${String(refreshError)}`
+          : `${successText}; state refresh failed, refresh manually: ${String(refreshError)}`);
+      }
+    } catch (importError) {
+      setError(String(importError));
+    } finally {
+      endLoading(loadingToken);
+      endActionBusy(actionToken);
+    }
+  };
+
+  const openExternalUrl = React.useCallback((url?: string | null) => {
+    if (!url) return;
+    window.setTimeout(() => {
+      void codexApi.openUrl(url).catch(() => {
+        setToast(lang === "zh" ? "打开浏览器失败" : "Failed to open browser");
+      });
+    }, 0);
+  }, [lang]);
+
+  const checkForUpdates = React.useCallback(async ({ quiet = false }: { quiet?: boolean } = {}) => {
+    setReleaseInfo({ status: "checking" });
+    try {
+      if (aboutInfo?.nativeUpdaterSupported !== false) {
+        const updaterResult = await appUpdater.check({ force: !quiet, timeout: 15_000 });
+        if (updaterResult === "available") {
+          const snapshot = appUpdater.getSnapshot();
+          const latestVersion = snapshot.latestVersion || "";
+          const releaseTag = latestVersion.startsWith("v") ? latestVersion : `v${latestVersion}`;
+          setReleaseInfo({
+            status: "ok",
+            latestVersion: releaseTag,
+            htmlUrl: `https://github.com/${FALLBACK_GITHUB_REPO}/releases/tag/${releaseTag}`,
+            hasUpdate: true,
+            updateMethod: "native",
+          });
+          if (quiet) {
+            setToast(lang === "zh" ? `发现新版本 ${releaseTag}，可在概览页查看` : `New version ${releaseTag} is available`);
+          } else {
+            setUpdatePromptOpen(true);
+          }
+          return;
+        }
+
+        if (updaterResult === "up-to-date") {
+          setReleaseInfo({
+            status: "ok",
+            latestVersion: aboutInfo?.appVersion,
+            htmlUrl: `https://github.com/${FALLBACK_GITHUB_REPO}/releases/latest`,
+            hasUpdate: false,
+          });
+          if (!quiet) setToast(lang === "zh" ? "当前已是最新版本" : "You are up to date");
+          return;
+        }
+      }
+
+      // Keep the existing lightweight release check as a manual-download fallback for
+      // bootstrap and portable builds that cannot use the native updater yet.
+      const update = await updatesApi.checkAppUpdate();
+      const message = update.hasUpdate
+        ? (lang === "zh" ? "发现新版本" : "Update available")
+        : (lang === "zh" ? "当前已是最新版本" : "You are up to date");
+      setReleaseInfo({
+        status: "ok",
+        latestVersion: update.latestVersion,
+        htmlUrl: update.htmlUrl,
+        hasUpdate: update.hasUpdate,
+        updateMethod: update.hasUpdate ? "download" : undefined,
+      });
+      if (update.hasUpdate) {
+        if (quiet) {
+          setToast(lang === "zh" ? `发现新版本 ${update.latestVersion}，可在概览页查看` : `New version ${update.latestVersion} is available`);
+        } else {
+          setUpdatePromptOpen(true);
+        }
+      } else if (!quiet) {
+        setToast(message);
+      }
+    } catch {
+      const message = quiet ? (lang === "zh" ? "自动检查失败" : "Auto check failed") : (lang === "zh" ? "检查失败" : "Check failed");
+      setReleaseInfo({
+        status: "error",
+      });
+      if (!quiet) setToast(message);
+    }
+  }, [aboutInfo?.appVersion, aboutInfo?.nativeUpdaterSupported, lang]);
+
+  React.useEffect(() => {
+    if (!state || !aboutInfo || autoUpdateCheckedRef.current) return;
+    autoUpdateCheckedRef.current = true;
+    void checkForUpdates({ quiet: true });
+  }, [aboutInfo, state, checkForUpdates]);
+
+  React.useEffect(() => {
+    if (!state || tab !== "instruction" || promptAutoRefreshAttemptedRef.current) return;
+    promptAutoRefreshAttemptedRef.current = true;
+    void refreshBuiltinPrompts({ quiet: true });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [state, tab]);
+
+  const beginSkillsMcpRequest = React.useCallback(() => ({
+    requestId: ++skillsMcpRequestRef.current,
+    configDir: configDir || null,
+    configDirKey: normalizedConfigDirForComparison(configDir),
+  }), [configDir]);
+
+  const isCurrentSkillsMcpRequest = React.useCallback((requestId: number, configDirKey: string) => (
+    requestId === skillsMcpRequestRef.current
+      && configDirKey === activeConfigDirKeyRef.current
+  ), []);
+
+  const loadSkillsMcp = React.useCallback(async ({ quiet = false }: { quiet?: boolean } = {}) => {
+    const request = beginSkillsMcpRequest();
+    const actionToken = quiet ? null : beginActionBusy("loadSkillsMcp");
+    if (!quiet) setError("");
+    try {
+      const result = await skillsMcpApi.getSkillsMcpState(request.configDir);
+      if (!isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) return;
+      skillsMcpLoadedRef.current = request.configDirKey;
+      setSkillsMcpState(result);
+    } catch (e) {
+      if (!quiet && isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) {
+        setError(String(e));
+      }
+    } finally {
+      if (actionToken !== null) endActionBusy(actionToken);
+    }
+  }, [beginActionBusy, beginSkillsMcpRequest, endActionBusy, isCurrentSkillsMcpRequest]);
+
+  React.useEffect(() => {
+    const configDirKey = normalizedConfigDirForComparison(configDir);
+    if (tab !== "skillsMcp") {
+      skillsMcpAutoLoadAttemptedRef.current = "";
+      return;
+    }
+    if (!state?.codexDir
+      || !configDirKey
+      || Boolean(actionBusy)
+      || skillsMcpAutoLoadAttemptedRef.current === configDirKey
+      || skillsMcpLoadedRef.current === configDirKey) return;
+    skillsMcpAutoLoadAttemptedRef.current = configDirKey;
+    void loadSkillsMcp();
+  }, [actionBusy, configDir, loadSkillsMcp, state?.codexDir, tab]);
+
+  const openImportExistingSkillsMcpPreview = async () => {
+    const request = beginSkillsMcpRequest();
+    const actionToken = beginActionBusy("previewExistingSkillsMcp");
+    setError("");
+    try {
+      const preview = await skillsMcpApi.previewExistingSkillsMcp(request.configDir);
+      if (!isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) return;
+      if (preview.skills.length + preview.mcpServers.length === 0) {
+        setSkillsMcpImportPreview(null);
+        setSkillsMcpImportOpen(false);
+        setToast(lang === "zh" ? "没有需要新导入的 Skills 或 MCP" : "No new Skills or MCP items to import");
+        return;
+      }
+      setSkillsMcpImportPreview(preview);
+      setSkillsMcpImportOpen(true);
+    } catch (e) {
+      if (isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) setError(String(e));
+    } finally {
+      endActionBusy(actionToken);
+    }
+  };
+
+  const importExistingSkillsMcp = async () => {
+    const request = beginSkillsMcpRequest();
+    const actionToken = beginActionBusy("importExistingSkillsMcp");
+    setError("");
+    try {
+      const result = await skillsMcpApi.importExistingSkillsMcp(request.configDir);
+      if (!isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) return;
+      skillsMcpLoadedRef.current = request.configDirKey;
+      setSkillsMcpState(result.state);
+      setSkillsMcpImportOpen(false);
+      setSkillsMcpImportPreview(null);
+      setToast(result.message);
+    } catch (e) {
+      if (isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) setError(String(e));
+    } finally {
+      endActionBusy(actionToken);
+    }
+  };
+
+  const checkSkillUpdatesAction = async () => {
+    const request = beginSkillsMcpRequest();
+    const actionToken = beginActionBusy("checkSkillUpdates");
+    setError("");
+    try {
+      const result = await skillsMcpApi.checkSkillUpdates(request.configDir);
+      if (!isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) return;
+      skillsMcpLoadedRef.current = request.configDirKey;
+      setSkillsMcpState(result);
+      setToast(lang === "zh" ? "Skills 更新状态已刷新" : "Skill update status refreshed");
+    } catch (e) {
+      if (isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) setError(String(e));
+    } finally {
+      endActionBusy(actionToken);
+    }
+  };
+
+  const toggleSkillEnabled = async (id: string, enabled: boolean) => {
+    const request = beginSkillsMcpRequest();
+    const actionToken = beginActionBusy(`skill:${id}`);
+    setError("");
+    try {
+      const result = await skillsMcpApi.toggleCodexSkill(request.configDir, id, enabled);
+      if (!isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) return;
+      skillsMcpLoadedRef.current = request.configDirKey;
+      setSkillsMcpState(result);
+      setToast(enabled ? (lang === "zh" ? "Skill 已启用" : "Skill enabled") : (lang === "zh" ? "Skill 已禁用" : "Skill disabled"));
+    } catch (e) {
+      if (isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) setError(String(e));
+    } finally {
+      endActionBusy(actionToken);
+    }
+  };
+
+  const toggleMcpEnabled = async (id: string, enabled: boolean) => {
+    const request = beginSkillsMcpRequest();
+    const actionToken = beginActionBusy(`mcp:${id}`);
+    setError("");
+    try {
+      const result = await skillsMcpApi.toggleCodexMcp(request.configDir, id, enabled);
+      if (!isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) return;
+      skillsMcpLoadedRef.current = request.configDirKey;
+      setSkillsMcpState(result);
+      setToast(enabled ? (lang === "zh" ? "MCP 已启用" : "MCP enabled") : (lang === "zh" ? "MCP 已禁用" : "MCP disabled"));
+    } catch (e) {
+      if (isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) setError(String(e));
+    } finally {
+      endActionBusy(actionToken);
+    }
+  };
+
+  const saveSkillsMcpNote = async (
+    itemKind: "skill" | "mcp",
+    itemId: string,
+    note: string,
+  ) => {
+    const requestConfigDir = configDir || null;
+    const requestConfigDirKey = normalizedConfigDirForComparison(configDir);
+    setError("");
+    try {
+      const result = await skillsMcpApi.updateSkillsMcpNote(
+        requestConfigDir,
+        itemKind,
+        itemId,
+        note,
+      );
+      if (requestConfigDirKey !== activeConfigDirKeyRef.current) {
+        throw new Error(lang === "zh" ? "CODEX_HOME 已切换，请重新保存备注" : "CODEX_HOME changed. Save the note again.");
+      }
+      setSkillsMcpState((current) => {
+        if (!current) return current;
+        if (result.itemKind === "skill") {
+          return {
+            ...current,
+            skills: current.skills.map((skill) => skill.id === result.itemId
+              ? { ...skill, note: result.note ?? null }
+              : skill),
+          };
+        }
+        return {
+          ...current,
+          mcpServers: current.mcpServers.map((server) => server.id === result.itemId
+            ? { ...server, note: result.note ?? null }
+            : server),
+        };
+      });
+    } catch (noteError) {
+      setError(String(noteError));
+      throw noteError;
+    }
+  };
+
+  const installSkillZipFile = async (file?: File | null) => {
+    if (!file) return;
+    if (!file.name.toLowerCase().endsWith(".zip")) {
+      setError(lang === "zh" ? "请选择 .zip 技能包" : "Please choose a .zip skill package");
+      return;
+    }
+    if (file.size > 20 * 1024 * 1024) {
+      setError(lang === "zh" ? "ZIP 技能包不能超过 20MB" : "Skill ZIP must be smaller than 20MB");
+      return;
+    }
+    const request = beginSkillsMcpRequest();
+    const actionToken = beginActionBusy("installSkillZip");
+    setError("");
+    try {
+      const bytes = Array.from(new Uint8Array(await file.arrayBuffer()));
+      if (!isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) return;
+      const result = await skillsMcpApi.installSkillZip(request.configDir, file.name, bytes);
+      if (!isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) return;
+      skillsMcpLoadedRef.current = request.configDirKey;
+      setSkillsMcpState(result.state);
+      setToast(result.message);
+    } catch (e) {
+      if (isCurrentSkillsMcpRequest(request.requestId, request.configDirKey)) setError(String(e));
+    } finally {
+      endActionBusy(actionToken);
+      if (skillZipImportRef.current) skillZipImportRef.current.value = "";
+    }
+  };
+
+  const officialAuthPlaceholder = '{\n  "OPENAI_API_KEY": null,\n  "auth_mode": "chatgpt",\n  "tokens": {\n    "access_token": "",\n    "refresh_token": "",\n    "id_token": ""\n  }\n}';
+
+  const openOfficialEdit = async (accountId?: string) => {
+    const requestId = ++officialDraftRequestRef.current;
+    const actionToken = beginActionBusy("loadOfficialDraft");
+    const liveIsOfficial = Boolean(state?.isOfficialProvider);
+    const fallbackModel = state?.model || "gpt-5.5";
+    const fallbackConfigText = liveIsOfficial && state?.configText?.trim()
+      ? state.configText
+      : "";
+    setEditingDetectedProvider(false);
+    setEditingOfficialAccountId(accountId || null);
+    setOfficialForm({
+      name: "",
+      model: fallbackModel,
+      authJson: liveIsOfficial && state?.authText ? state.authText : officialAuthPlaceholder,
+      configText: fallbackConfigText,
+    });
+    setProviderMode("official");
+    setError("");
+    try {
+      if (accountId) {
+        const account = await officialAccountsApi.getOfficialAccount(configDir || null, accountId);
+        if (requestId !== officialDraftRequestRef.current) return;
+        setOfficialForm({
+          name: account.name,
+          model: account.model || fallbackModel,
+          authJson: account.authJson,
+          configText: account.configText,
+        });
+      } else {
+        const draft = await officialAccountsApi.getOfficialConfigDraft(configDir || null);
+        if (requestId !== officialDraftRequestRef.current) return;
+        if (draft) {
+          setOfficialForm({
+            name: "",
+            model: draft.model || fallbackModel,
+            authJson: draft.authJson,
+            configText: draft.configText || fallbackConfigText,
+          });
+        }
+      }
+    } catch (editError) {
+      if (requestId === officialDraftRequestRef.current) setError(String(editError));
+    } finally {
+      endActionBusy(actionToken);
+    }
+  };
+
+  const saveOfficialConfig = () => {
+    if (editingOfficialAccountId && !officialForm.name.trim()) {
+      setError(lang === "zh" ? "账号名称不能为空" : "Account name is required");
+      return;
+    }
+    const before = state;
+    call(
+      () => editingOfficialAccountId
+        ? officialAccountsApi.updateOfficialAccount({
+          configDir: configDir || null,
+          accountId: editingOfficialAccountId,
+          name: officialForm.name,
+          model: officialForm.model,
+          authJson: officialForm.authJson,
+          configText: officialForm.configText,
+        })
+        : officialAccountsApi.saveOfficialConfig({
+          configDir: configDir || null,
+          model: officialForm.model,
+          authJson: officialForm.authJson,
+          configText: officialForm.configText,
+        }),
+      (result) => {
+        handleConfigActionResult(before, result);
+        setEditingOfficialAccountId(null);
+        setProviderMode("list");
+      },
+    );
+  };
+
+  const openAddProvider = () => {
+    const liveToml = state?.configText?.trim() || "";
+    const next = {
+      ...blankProviderForm,
+      model: state?.model?.trim() || blankProviderForm.model,
+      wireApi: currentProvider?.wireApi?.trim() || blankProviderForm.wireApi,
+      requiresOpenaiAuth: currentProvider?.requiresOpenaiAuth ?? blankProviderForm.requiresOpenaiAuth,
+      tomlConfig: liveToml,
+    };
+    resetAvailableProviderModels();
+    setEditingProviderId(null);
+    setCopyingProviderId(null);
+    setEditingDetectedProvider(false);
+    setProviderForm(next);
+    setProviderTomlDraft(liveToml || buildProviderTomlPreview(next));
+    setProviderTomlDirty(false);
+    setProviderMode("form");
+  };
+
+  const openEditProvider = (provider: SavedProvider) => {
+    resetAvailableProviderModels();
+    setEditingProviderId(provider.id);
+    setCopyingProviderId(null);
+    setEditingDetectedProvider(false);
+    setProviderForm(provider);
+    setProviderTomlDraft(provider.tomlConfig?.trim() || buildProviderTomlPreview(provider));
+    setProviderTomlDirty(false);
+    setProviderMode("form");
+  };
+
+  const openCopyProvider = (provider: SavedProvider) => {
+    resetAvailableProviderModels();
+    const providerName = `${provider.providerName} ${lang === "zh" ? "副本" : "Copy"}`;
+    const id = uniqueId(customProviderId(providerName), savedProviders.map((item) => item.id));
+    const copy = {
+      ...provider,
+      id,
+      providerName,
+      apiKey: provider.apiKey || "",
+      tomlConfig: provider.tomlConfig || "",
+    };
+    setEditingProviderId(null);
+    setCopyingProviderId(provider.id);
+    setEditingDetectedProvider(false);
+    setProviderForm(copy);
+    setProviderTomlDraft(copy.tomlConfig.trim() || buildProviderTomlPreview(copy));
+    setProviderTomlDirty(false);
+    setProviderMode("form");
+  };
+
+  const openEditDetectedProvider = (provider: { id: string; providerName: string; baseUrl: string; model: string; apiKey?: string; wireApi: string; requiresOpenaiAuth: boolean }) => {
+    resetAvailableProviderModels();
+    const id = uniqueId(
+      customProviderId(provider.providerName || provider.baseUrl),
+      savedProviders.map((item) => item.id),
+    );
+    setEditingProviderId(id);
+    setCopyingProviderId(null);
+    setEditingDetectedProvider(true);
+    const next = {
+      id,
+      providerName: provider.providerName,
+      baseUrl: provider.baseUrl,
+      model: provider.model,
+      apiKey: provider.apiKey || "",
+      tomlConfig: state?.configText?.trim() || "",
+      wireApi: provider.wireApi || "responses",
+      requiresOpenaiAuth: provider.requiresOpenaiAuth,
+    };
+    setProviderForm(next);
+    setProviderTomlDraft(next.tomlConfig || buildProviderTomlPreview(next));
+    setProviderTomlDirty(false);
+    setProviderMode("form");
+  };
+
+  const removeProvider = async (id: string, isCurrent: boolean) => {
+    const before = state;
+    const loadingToken = beginLoading();
+    setError("");
+    try {
+      let switched: ActionResult | null = null;
+      if (isCurrent) {
+        switched = await officialAccountsApi.switchOfficialProvider(configDir || null);
+        persistActiveProviderId("");
+        setActiveProviderId("");
+      }
+      await providersApi.deleteSavedProvider(id, configDir || null);
+      const providerList = await providersApi.listSavedProviders();
+      setSavedProviders(providerList);
+      if (switched) handleConfigActionResult(before, switched);
+      setToast(lang === "zh" ? "供应商已删除" : "Provider deleted");
+      return true;
+    } catch (e) {
+      setError(String(e));
+      return false;
+    } finally {
+      endLoading(loadingToken);
+    }
+  };
+
+  const checkSessions = async () => {
+    sessionLoadRequestRef.current += 1;
+    const actionToken = beginActionBusy("checkSessions");
+    setSessionStatus(null);
+    await call(
+      () => sessionsApi.getSessionSyncStatus(configDir || null),
+      (status) => {
+        setSessionStatus(status);
+        if (!status.scanComplete) {
+          setToast(status.scanFailures[0] || (lang === "zh" ? "无法确认会话同步状态" : "Unable to verify session sync status"));
+          return;
+        }
+        const hasMismatches = Boolean(status.needsSync);
+        const syncCount = sessionMismatchCount(status);
+        setToast(hasMismatches
+          ? (lang === "zh" ? `有 ${syncCount} 条会话需要同步` : `${syncCount} session(s) need syncing`)
+          : (lang === "zh" ? "全部会话已同步" : "All sessions are synced"));
+      },
+    );
+    endActionBusy(actionToken);
+  };
+
+  React.useEffect(() => {
+    if (tab !== "sessions" || refreshing || !state?.codexDir) return;
+    const loadKey = state.codexDir;
+    if (sessionAutoLoadKeyRef.current === loadKey) return;
+    const requestId = ++sessionLoadRequestRef.current;
+    const actionToken = beginActionBusy("checkSessions");
+    sessionAutoLoadKeyRef.current = loadKey;
+    void sessionsApi.getSessionSyncStatus(loadKey)
+      .then((status) => {
+        if (requestId === sessionLoadRequestRef.current) setSessionStatus(status);
+      })
+      .catch((sessionError) => {
+        if (requestId !== sessionLoadRequestRef.current) return;
+        sessionAutoLoadKeyRef.current = "";
+        setError(String(sessionError));
+      })
+      .finally(() => {
+        endActionBusy(actionToken);
+      });
+  }, [beginActionBusy, endActionBusy, refreshing, state?.codexDir, tab]);
+
+  const syncSessions = async () => {
+    sessionLoadRequestRef.current += 1;
+    const actionToken = beginActionBusy("syncSessions");
+    await call(
+      () => sessionsApi.syncSessionsProvider(configDir || null),
+      (result) => {
+        setSessionStatus(result.status);
+        setSelectedSessionIds([]);
+        if (result.desktopLifecycleWarning) {
+          setToast(result.desktopLifecycleWarning);
+        } else if (!result.status.scanComplete) {
+          setToast(result.status.scanFailures[0] || (lang === "zh" ? "无法确认会话同步状态" : "Unable to verify session sync status"));
+        } else if (result.status.needsSync) {
+          const remaining = sessionMismatchCount(result.status);
+          setToast(lang === "zh"
+            ? `已同步可写的会话索引，仍有 ${remaining} 条需要重试；聊天内容未改动`
+            : `Writable session indexes were synced; ${remaining} still need retrying. Chat content was not changed.`);
+        } else {
+          setToast(lang === "zh"
+            ? "会话索引已全部同步，聊天内容未改动"
+            : "All session indexes are synced. Chat content was not changed.");
+        }
+      },
+    );
+    endActionBusy(actionToken);
+  };
+
+  const toggleSessionSelected = (id: string) => {
+    setSelectedSessionIds((ids) => ids.includes(id) ? ids.filter((item) => item !== id) : [...ids, id]);
+  };
+
+  const setSessionGroupSelected = (sessions: SessionPreview[], checked: boolean) => {
+    const groupIds = new Set(sessions.map((item) => item.id));
+    setSelectedSessionIds((ids) => {
+      const next = new Set(ids);
+      if (checked) groupIds.forEach((id) => next.add(id));
+      else groupIds.forEach((id) => next.delete(id));
+      return Array.from(next);
+    });
+  };
+
+  const closeSessionDeleteConfirm = () => {
+    if (!sessionDeleteBusy) {
+      setSessionDeleteConfirmOpen(false);
+      setSessionDeleteSafetyConfirmed(false);
+    }
+  };
+
+  const deleteSelectedSessions = async () => {
+    if (!selectedSessionIds.length || sessionDeleteBusy || !sessionDeleteSafetyConfirmed) return;
+    setSessionDeleteBusy(true);
+    setToast("");
+    setError("");
+    try {
+      const result = await sessionsApi.deleteCodexSessions(configDir || null, selectedSessionIds);
+      setSessionStatus(result.status);
+      const remainingIds = new Set(result.status.sessions.map((item) => item.id));
+      setSelectedSessionIds((ids) => ids.filter((id) => remainingIds.has(id)));
+      setSessionDeleteConfirmOpen(false);
+      setSessionDeleteSafetyConfirmed(false);
+      const hasPartialFailure = result.failedSessions > 0 || Boolean(result.failureMessage);
+      if (hasPartialFailure) {
+        setError(result.failureMessage || (lang === "zh"
+          ? `${result.failedSessions} 个会话删除失败，请关闭其他 Codex 窗口或 CLI 后重试。`
+          : `${result.failedSessions} session deletion(s) failed. Close other Codex windows or CLIs and retry.`));
+      } else if (result.desktopLifecycleWarning) {
+        setToast(result.desktopLifecycleWarning);
+      } else {
+        setToast(lang === "zh"
+          ? `已永久删除 ${result.deletedSessions} 条会话，并清理数据库、Rollout 与关联历史`
+          : `Permanently deleted ${result.deletedSessions} session(s) and cleaned database, rollout, and related history data`);
+      }
+    } catch (e) {
+      setError(String(e));
+    } finally {
+      setSessionDeleteBusy(false);
+    }
+  };
+
+  const closeStartupWizard = () => {
+    markStartupWizardSeen();
+    setStartupClosing(true);
+    window.setTimeout(() => {
+      setStartupWizardOpen(false);
+      setStartupClosing(false);
+    }, 260);
+  };
+
+  const changeTab = (nextTab: Tab) => {
+    if (nextTab !== "provider") {
+      officialDraftRequestRef.current += 1;
+      clearActionBusy("loadOfficialDraft");
+    }
+    setTab(nextTab);
+  };
+
+  return (
+    <AppShell
+      activeTab={tab}
+      onTabChange={changeTab}
+      lang={lang}
+      theme={theme}
+      onToggleTheme={toggleTheme}
+      codexVersion={aboutInfo?.codexVersion
+        || (aboutLoading ? (lang === "zh" ? "正在检测..." : "Detecting...") : undefined)}
+      appVersion={aboutInfo?.appVersion}
+      hasUpdate={Boolean(releaseInfo.hasUpdate)}
+      updatePhase={updater.state.phase}
+      onOpenUpdate={() => setUpdatePromptOpen(true)}
+      isMacRuntime={isMacRuntime}
+      contentClassName={cx(
+        tab === "sessions" && "cx-app-content--sessions",
+        (
+          (tab === "provider" && providerMode === "list")
+          || tab === "skillsMcp"
+        ) && "cx-app-content--fixed",
+        skillsMcpImportOpen && Boolean(skillsMcpImportPreview) && "cx-app-content--modal-locked",
+      )}
+    >
+      <AppToast
+        lang={lang}
+        message={toast}
+        error={error}
+        loading={Boolean(providerTestingId || providerModelsLoading) && Boolean(toast)}
+        onDismissMessage={() => setToast("")}
+        onDismissError={() => setError("")}
+      />
+      <UpdateDialog
+        open={updatePromptOpen && Boolean(releaseInfo.hasUpdate)}
+        lang={lang}
+        state={releaseInfo.updateMethod === "native" ? updater.state : undefined}
+        currentVersion={aboutInfo?.appVersion}
+        latestVersion={releaseInfo.latestVersion}
+        onClose={() => setUpdatePromptOpen(false)}
+        onUpdate={releaseInfo.updateMethod === "native" ? updater.downloadAndInstall : undefined}
+        onRetry={releaseInfo.updateMethod === "native" ? updater.retry : undefined}
+        onRestart={releaseInfo.updateMethod === "native" ? updater.restart : undefined}
+        onDownload={() => {
+          setUpdatePromptOpen(false);
+          openExternalUrl(releaseInfo.htmlUrl);
+        }}
+      />
+      <StartupWizardDialog
+        open={startupWizardOpen}
+        closing={startupClosing}
+        lang={lang}
+        diagnostics={startupDiagnostics}
+        configDir={configDirDraft}
+        loading={loading || refreshing}
+        onConfigDirChange={setConfigDirDraft}
+        onRecheck={() => refresh(true)}
+        onSkip={closeStartupWizard}
+        onOpenSettings={() => {
+          changeTab("settings");
+          closeStartupWizard();
+        }}
+        onEnter={closeStartupWizard}
+      />
+
+      <PageTransition pageKey={tab}>
+            {!state && tab !== "dashboard" && tab !== "settings" && tab !== "about" && (
+              <CodexStateLoading lang={lang} loading={refreshing} />
+            )}
+
+            {tab === "dashboard" && (
+              <OverviewPage
+                lang={lang}
+                ready={Boolean(state)}
+                model={state?.model}
+                configDir={configDirDraft}
+                resolvedCodexDir={state?.codexDir || ""}
+                configExists={Boolean(state?.configExists)}
+                providerLabel={currentProvider?.name || state?.modelProvider}
+                instructionEnabled={Boolean(state?.instructionEnabled)}
+                authExists={Boolean(state?.authExists)}
+                officialAuthAvailable={Boolean(state?.officialAuthAvailable)}
+                configPath={state?.configPath}
+                modelProvider={state?.modelProvider}
+                instructionPath={state
+                  ? (state.instructionInjectionMode === "append"
+                    ? `${state.agentsPath} (${lang === "zh" ? "追加模式" : "append"})`
+                    : state.instructionFile)
+                  : null}
+                loading={loading || refreshing}
+                hasUpdate={Boolean(releaseInfo.status === "ok" && releaseInfo.hasUpdate)}
+                latestVersion={releaseInfo.latestVersion}
+                onConfigDirChange={setConfigDirDraft}
+                onRefresh={() => refresh(false)}
+                onOpenUpdate={() => setUpdatePromptOpen(true)}
+              />
+            )}
+
+            {state && tab === "provider" && (
+              <ProvidersPage
+                lang={lang}
+                copy={getProviderPageCopy(lang)}
+                mode={providerMode}
+                providerRows={providerPageRows}
+                loading={loading}
+                testingId={providerTestingId}
+                actionBusy={actionBusy}
+                isRestartingCodex={isRestartingCodex}
+                editingProviderId={editingProviderId || (editingDetectedProvider ? providerForm.id : null)}
+                editingOfficialAccountId={editingOfficialAccountId}
+                providerForm={{
+                  apiKey: providerForm.apiKey || "",
+                  baseUrl: providerForm.baseUrl,
+                  providerName: providerForm.providerName,
+                  model: providerForm.model,
+                  wireApi: providerForm.wireApi,
+                  requiresOpenaiAuth: providerForm.requiresOpenaiAuth,
+                }}
+                officialForm={officialForm}
+                officialAuthRef={officialAuthEditorRef}
+                officialTomlRef={officialTomlEditorRef}
+                officialInfo={{
+                  officialUrl: "https://chatgpt.com/codex",
+                  authPath: state.authPath,
+                  current: state.isOfficialProvider
+                    ? officialAccounts.find((account) => account.selected)?.name || "OpenAI Official"
+                    : state.modelProvider,
+                }}
+                providerAuthPreview={<JsonPreview text={providerAuthPreview} />}
+                providerTomlDraft={providerTomlDraft}
+                providerTomlRef={providerTomlEditorRef}
+                apiKeyVisible={providerApiKeyVisible}
+                availableModels={availableProviderModels.map((model) => model.id)}
+                fetchingModels={providerModelsLoading}
+                onImportCcSwitch={importFromCcSwitch}
+                onRestartCodex={() => void restartCodex()}
+                onAddProvider={openAddProvider}
+                onPrepareNewOfficialAccount={prepareNewOfficialAccount}
+                showSaveCurrentOfficial={Boolean(state.isOfficialProvider && !officialAccounts.some((account) => account.selected))}
+                onCaptureCurrentOfficial={captureCurrentOfficialAccount}
+                onLoadCcSwitchOfficial={() => void loadCcSwitchOfficial()}
+                onEnableProvider={(row) => {
+                  if (row.source === "official") {
+                    if (row.officialAccountId) switchOfficialAccount(row.officialAccountId);
+                    else switchOfficialProvider();
+                    return;
+                  }
+                  const local = findLocalProviderForRow(row);
+                  switchProvider(local || {
+                    id: customProviderId(row.providerName),
+                    providerName: row.providerName,
+                    baseUrl: row.baseUrl,
+                    model: row.model,
+                    apiKey: row.apiKey || "",
+                    tomlConfig: "",
+                    wireApi: row.wireApi,
+                    requiresOpenaiAuth: row.requiresOpenaiAuth,
+                  });
+                }}
+                onTestProvider={(row) => {
+                  const local = findLocalProviderForRow(row);
+                  void testProvider(row.testingKey || `${row.source}-${row.id}`, row.baseUrl, local?.apiKey || row.apiKey || null);
+                }}
+                onEditProvider={(row) => {
+                  if (row.source === "official") {
+                    void openOfficialEdit(row.officialAccountId);
+                    return;
+                  }
+                  const local = findLocalProviderForRow(row);
+                  if (local) openEditProvider(local);
+                  else if (row.source === "detected") openEditDetectedProvider(row);
+                }}
+                onCopyProvider={(row) => {
+                  const local = findLocalProviderForRow(row);
+                  if (local) openCopyProvider(local);
+                }}
+                onDeleteProvider={(row) => {
+                  if (row.officialAccountId) return deleteOfficialAccount(row.officialAccountId);
+                  const local = findLocalProviderForRow(row);
+                  return local ? removeProvider(local.id, row.isCurrent) : Promise.resolve(false);
+                }}
+                onRestoreOfficial={restoreOfficialProvider}
+                onResetOfficial={resetOfficialProvider}
+                onCancelMode={() => {
+                  officialDraftRequestRef.current += 1;
+                  setProviderMode("list");
+                  setEditingOfficialAccountId(null);
+                  setCopyingProviderId(null);
+                  setEditingDetectedProvider(false);
+                  setProviderTomlDirty(false);
+                }}
+                onOfficialNameChange={(value) => setOfficialForm((current) => ({ ...current, name: value }))}
+                onOfficialModelChange={(value) => setOfficialForm((current) => ({ ...current, model: value }))}
+                onOfficialAuthChange={(value) => setOfficialForm((current) => ({ ...current, authJson: value }))}
+                onOfficialConfigChange={(value) => setOfficialForm((current) => ({ ...current, configText: value }))}
+                onSaveOfficial={saveOfficialConfig}
+                onApiKeyChange={(value) => {
+                  resetAvailableProviderModels();
+                  setProviderForm((current) => ({ ...current, apiKey: value }));
+                }}
+                onBaseUrlChange={(value) => {
+                  resetAvailableProviderModels();
+                  setProviderForm((current) => ({ ...current, baseUrl: value }));
+                }}
+                onProviderNameChange={(value) => setProviderForm((current) => ({
+                  ...current,
+                  providerName: value,
+                  id: (editingProviderId || copyingProviderId) ? current.id : customProviderId(value),
+                }))}
+                onProviderModelChange={(value) => setProviderForm((current) => ({ ...current, model: value }))}
+                onFetchModels={() => void fetchProviderModels()}
+                onWireApiChange={(value) => setProviderForm((current) => ({ ...current, wireApi: value }))}
+                onRequiresAuthChange={(value) => setProviderForm((current) => ({ ...current, requiresOpenaiAuth: value }))}
+                onToggleApiKeyVisibility={() => setProviderApiKeyVisible((value) => !value)}
+                onProviderTomlDraftChange={(value) => {
+                  providerDraftRequestRef.current += 1;
+                  setProviderTomlDraft(value);
+                  setProviderTomlDirty(true);
+                }}
+                onResetProviderToml={() => {
+                  providerDraftRequestRef.current += 1;
+                  setProviderTomlDraft(
+                    providerForm.tomlConfig?.trim()
+                    || state?.configText?.trim()
+                    || providerTomlPreview,
+                  );
+                  setProviderTomlDirty(false);
+                  setProviderDraftRefreshToken((token) => token + 1);
+                }}
+                onSaveProvider={saveProviderConfig}
+              />
+            )}
+
+            {state && (tab === "sessions" || visitedTabs.has("sessions")) && (
+              <SessionManagementPage
+                active={tab === "sessions"}
+                lang={lang}
+                sessionStatus={sessionStatus}
+                sessionHasMismatches={sessionHasMismatches}
+                sessionSyncCount={sessionSyncCount}
+                sessionTargetLabel={sessionTargetLabel}
+                sessionVisibleTotal={sessionVisibleTotal}
+                sessionPreviewTruncated={sessionPreviewTruncated}
+                visibleSessions={visibleSessions}
+                filteredSessions={filteredSessions}
+                allSessionsByCwd={allSessionsByCwd}
+                groupedSessions={groupedSessions}
+                selectedSessionIds={selectedSessionIds}
+                selectedSessionSet={selectedSessionSet}
+                selectedSessions={selectedSessions}
+                sessionQuery={sessionQuery}
+                sessionGroupByCwd={sessionGroupByCwd}
+                showInternalSessions={showInternalSessions}
+                loading={loading}
+                actionBusy={actionBusy}
+                sessionDeleteConfirmOpen={sessionDeleteConfirmOpen}
+                sessionDeleteBusy={sessionDeleteBusy}
+                sessionDeleteSafetyConfirmed={sessionDeleteSafetyConfirmed}
+                onCheckSessions={checkSessions}
+                onSyncSessions={syncSessions}
+                onSessionQueryChange={(value) => {
+                  setSessionQuery(value);
+                  setSelectedSessionIds([]);
+                  setSessionDeleteConfirmOpen(false);
+                }}
+                onSessionGroupByCwdChange={setSessionGroupByCwd}
+                onShowInternalSessionsChange={(checked) => {
+                  setShowInternalSessions(checked);
+                  setSelectedSessionIds([]);
+                  setSessionDeleteConfirmOpen(false);
+                }}
+                onOpenDeleteConfirm={() => {
+                  setSessionDeleteSafetyConfirmed(false);
+                  setSessionDeleteConfirmOpen(true);
+                }}
+                onToggleSessionSelected={toggleSessionSelected}
+                onSetSessionGroupSelected={setSessionGroupSelected}
+                onCloseDeleteConfirm={closeSessionDeleteConfirm}
+                onDeleteSelectedSessions={deleteSelectedSessions}
+                onDeleteSafetyConfirmedChange={setSessionDeleteSafetyConfirmed}
+              />
+            )}
+
+            {state && (tab === "skillsMcp" || visitedTabs.has("skillsMcp")) && (
+              <SkillsMcpPage
+                lang={lang}
+                state={skillsMcpState}
+                activeTab={skillsMcpTab}
+                actionBusy={actionBusy}
+                importOpen={skillsMcpImportOpen}
+                importPreview={skillsMcpImportPreview}
+                zipInputRef={skillZipImportRef}
+                className={tab !== "skillsMcp" ? "page-pane-hidden" : undefined}
+                onTabChange={setSkillsMcpTab}
+                onLoad={loadSkillsMcp}
+                onOpenImportPreview={openImportExistingSkillsMcpPreview}
+                onCloseImportPreview={() => setSkillsMcpImportOpen(false)}
+                onConfirmImport={importExistingSkillsMcp}
+                onInstallZip={installSkillZipFile}
+                onCheckUpdates={checkSkillUpdatesAction}
+                onToggleSkill={toggleSkillEnabled}
+                onToggleMcp={toggleMcpEnabled}
+                onSaveNote={saveSkillsMcpNote}
+              />
+            )}
+
+            {state && tab === "instruction" && (
+              <PromptsPage
+                lang={lang}
+                instructionMode={instructionMode}
+                promptForm={promptForm}
+                editingPromptId={editingPromptId}
+                editingBuiltinPrompt={editingBuiltinPrompt}
+                loading={loading}
+                actionBusy={actionBusy}
+                promptSyncing={promptSyncing}
+                promptCatalogReady={promptCatalogReady}
+                promptImportRef={promptImportRef}
+                promptInjectionMode={promptInjectionMode}
+                promptModeHelpOpen={promptModeHelpOpen}
+                promptModeHelpRef={promptModeHelpRef}
+                instructionEnabled={state.instructionEnabled}
+                activeInstructionTitle={activeInstructionTitle}
+                activeInjectionMode={state.instructionInjectionMode}
+                instructionTemplates={instructionTemplates}
+                builtinPromptStatuses={builtinPromptStatus}
+                activeBuiltinTemplateId={activeBuiltinTemplateId}
+                orphanedBuiltinPrompt={missingActiveBuiltinTemplateId ? {
+                  id: missingActiveBuiltinTemplateId,
+                  title: activeInstructionTitle,
+                  description: lang === "zh"
+                    ? "该模板已从在线目录移除，当前配置仍在使用。"
+                    : "This template was removed online but is still active.",
+                } : null}
+                savedPrompts={savedPrompts}
+                managedSavedPromptId={state.instructionTemplateKey?.startsWith("saved:")
+                  ? state.instructionTemplateKey.slice("saved:".length)
+                  : null}
+                preservedSavedPromptFilename={state.instructionInjectionMode === "append" ? currentInstructionFilename : null}
+                externalPrompt={state.instructionFile
+                  && currentInstructionId === "custom"
+                  && !savedPrompts.some((prompt) => currentInstructionFilename === prompt.filename)
+                  && !(missingActiveBuiltinTemplateId && state.instructionInjectionMode !== "append")
+                  ? {
+                    title: lang === "zh" ? "用户原有指令提示词" : "Existing user prompt",
+                    description: state.instructionInjectionMode === "append"
+                      ? (lang === "zh"
+                        ? "追加模式已保留这份外部提示词，并同时加载 Codex-X 的 AGENTS.md 区块。"
+                        : "Append mode preserves this external prompt alongside the Codex-X AGENTS.md block.")
+                      : (lang === "zh"
+                        ? "当前使用的是非 Codex-X 管理的外部提示词。"
+                        : "This external prompt is not managed by Codex-X."),
+                    filename: currentInstructionFilename,
+                  }
+                  : null}
+                onSyncBuiltinPrompts={() => refreshBuiltinPrompts()}
+                onImportPrompt={importPromptMd}
+                onAddPrompt={openAddPrompt}
+                onInstructionModeChange={setInstructionMode}
+                onPromptInjectionModeChange={setPromptInjectionMode}
+                onTogglePromptModeHelp={() => setPromptModeHelpOpen((open) => !open)}
+                onEnableBuiltinPrompt={switchInstructionTemplate}
+                onDisableInstruction={disableInstruction}
+                onEnableSavedPrompt={enableSavedPrompt}
+                onDisableExternalPrompt={disableExternalInstruction}
+                onEditPrompt={openEditPrompt}
+                onEditBuiltinPrompt={openEditBuiltinPrompt}
+                onDeletePrompt={removeSavedPrompt}
+                onPromptFormFieldChange={(field, value) => setPromptForm((current) => ({
+                  ...current,
+                  [field]: value,
+                  ...(field === "title" ? { id: editingPromptId || providerId(value) } : {}),
+                }))}
+                onSavePrompt={savePromptOnly}
+              />
+            )}
+
+            {state && tab === "toml" && (
+              <TomlConfigPage
+                eyebrow="~/.codex/config.toml"
+                title={t.toml.title}
+                description={t.toml.desc}
+                loaded={state.configExists ? t.toml.loaded : t.dashboard.missing}
+                isLoaded={state.configExists}
+                preview={<TomlPreview text={state.configText || t.toml.missingText} />}
+              />
+            )}
+
+            {tab === "about" && (
+              <AboutPage
+                copy={{
+                  eyebrow: "About",
+                  title: lang === "zh" ? "关于 Codex-X" : "About Codex-X",
+                  appVersionLabel: `Codex-X ${lang === "zh" ? "版本" : "Version"}`,
+                  codexVersionLabel: `Codex CLI ${lang === "zh" ? "版本" : "Version"}`,
+                  codexHomeLabel: "CODEX_HOME",
+                  projectLabel: lang === "zh" ? "项目地址" : "Project",
+                  openProjectLabel: lang === "zh" ? "打开项目主页" : "Open project",
+                  openIssuesLabel: lang === "zh" ? "反馈问题" : "Issues",
+                  releasesEyebrow: "GitHub Releases",
+                  releasesTitle: lang === "zh" ? "更新检查" : "Update check",
+                  releaseStatusLabel: lang === "zh" ? "状态" : "Status",
+                  latestVersionLabel: lang === "zh" ? "最新版本" : "Latest version",
+                  checkUpdateLabel: lang === "zh" ? "检查更新" : "Check updates",
+                  openReleasesLabel: lang === "zh" ? "打开下载页" : "Open releases",
+                }}
+                appVersion={aboutInfo?.appVersion || (aboutLoading ? (lang === "zh" ? "正在检测" : "Detecting") : "-")}
+                codexVersion={aboutInfo?.codexVersion || (aboutLoading
+                  ? (lang === "zh" ? "正在检测..." : "Detecting...")
+                  : (lang === "zh" ? "未检测到" : "Not detected"))}
+                codexHome={aboutInfo?.codexDir || state?.codexDir || configDir || "~/.codex"}
+                projectUrl={aboutInfo?.projectUrl || `https://github.com/${FALLBACK_GITHUB_REPO}`}
+                release={{
+                  status: releaseStatusLabel,
+                  latestVersion: releaseInfo.latestVersion || "-",
+                  tone: releaseInfo.status === "error"
+                    ? "error"
+                    : releaseInfo.hasUpdate
+                      ? "warning"
+                      : releaseInfo.status === "ok"
+                        ? "success"
+                        : "neutral",
+                  checking: releaseInfo.status === "checking"
+                    || updater.state.phase === "downloading"
+                    || updater.state.phase === "installing",
+                  canOpenReleases: Boolean(releaseInfo.htmlUrl),
+                }}
+                onOpenProject={() => openExternalUrl(aboutInfo?.projectUrl || `https://github.com/${FALLBACK_GITHUB_REPO}`)}
+                onOpenIssues={() => openExternalUrl(`${aboutInfo?.projectUrl || `https://github.com/${FALLBACK_GITHUB_REPO}`}/issues`)}
+                onCheckUpdate={() => void checkForUpdates()}
+                onOpenReleases={() => openExternalUrl(releaseInfo.htmlUrl)}
+              />
+            )}
+
+            {tab === "settings" && (
+              <SettingsPage
+                lang={lang}
+                copy={{
+                  eyebrow: "Settings",
+                  title: t.settings.title,
+                  languageTitle: t.settings.language,
+                  languageDescription: t.settings.languageDesc,
+                  chineseLabel: t.settings.zh,
+                  englishLabel: t.settings.en,
+                  productTitle: t.settings.productName,
+                  productDescription: t.settings.productDesc,
+                  productValue: "Codex-X",
+                  recheckTitle: lang === "zh" ? "首次启动向导" : "First-run wizard",
+                  recheckDescription: lang === "zh"
+                    ? "重新检测 CODEX_HOME、config.toml、auth.json 和 SQLite 会话库。"
+                    : "Recheck CODEX_HOME, config.toml, auth.json and SQLite session stores.",
+                  recheckLabel: lang === "zh" ? "重新检测" : "Recheck",
+                  autoRestartTitle: lang === "zh"
+                    ? "配置变更后自动重启 Codex"
+                    : "Automatically restart Codex after configuration changes",
+                  autoRestartDescription: lang === "zh"
+                    ? "应用当前供应商或影响 Codex 运行配置后，自动重新启动 Codex Desktop。可能中断当前正在进行的会话。"
+                    : "Restart Codex Desktop after applying configuration changes. This may interrupt an active session.",
+                }}
+                onLanguageChange={setLang}
+                autoRestartChecked={autoRestartCodex}
+                onAutoRestartChange={setAutoRestartCodex}
+                recheckBusy={loading || refreshing}
+                onRecheck={() => {
+                  clearStartupWizardSeen();
+                  setStartupWizardOpen(true);
+                  refresh(true);
+                }}
+              />
+            )}
+      </PageTransition>
+    </AppShell>
+  );
+}
