+++ apps/desktop/src/features/sessions/types.ts
@@ -1,0 +1,34 @@
+export type SessionPreview = {
+  id: string;
+  title: string;
+  modelProvider?: string | null;
+  model?: string | null;
+  cwd?: string | null;
+  rolloutPath?: string | null;
+  updatedAtMs?: number | null;
+  archived: boolean;
+  hasUserEvent: boolean;
+  isSubagent: boolean;
+  needsSync: boolean;
+};
+
+export type SessionSyncStatus = {
+  codexDir: string;
+  targetProvider: string;
+  rolloutFiles: number;
+  sessionMetaCount: number;
+  mismatchedRollouts: number;
+  mismatchedSessionMeta: number;
+  sqliteDbs: number;
+  sqliteThreads: number;
+  topLevelThreads: number;
+  subagentThreads: number;
+  mismatchedThreads: number;
+  mismatchedSessions: number;
+  needsSync: boolean;
+  scanComplete: boolean;
+  scanFailures: string[];
+  backupDir?: string | null;
+  warnings: string[];
+  sessions: SessionPreview[];
+};
