+++ apps/desktop/src/features/updates/api.ts
@@ -1,0 +1,6 @@
+import { invokeCommand } from "../../shared/api/tauri";
+import type { AppUpdateInfo } from "../../types";
+
+export function checkAppUpdate() {
+  return invokeCommand<AppUpdateInfo>("check_app_update");
+}
