+++ apps/desktop/src/shared/api/tauri.ts
@@ -1,0 +1,8 @@
+import { invoke } from "@tauri-apps/api/core";
+
+export function invokeCommand<TResult>(
+  command: string,
+  args?: Record<string, unknown>,
+): Promise<TResult> {
+  return invoke<TResult>(command, args);
+}
